First large dagster rearchitecture
1) Introduce the notion of Source definitions in addition to inputs.
A SourceDefinition represents a method of instantiating an input
into a transform that is not just flowing it from an upstream
dependency e.g. loading a file or just laoding a bag of key-value pairs

2) Introduce the notion of materializations. These are the mirror
images of sources, but for the output. A given output definition knows
zero or more ways to materialize itself.

3) Strongly-typed interfaces to configure the execution and
materializations of pipelines.