Move config class creationt to a function
With the click api for make_pass_decorator you have a
invoke it on a created class. My general preference
is to avoid programmatic manipulation of classes in
global scope once they have been created. I'm not married
to this solution but I want class creation and any required
maninpulation grouped together somehow.