Change inputs to be scoped by solid- rather than pipeline-scoped
Previous to this PR input names were globally scope to a pipeline.

This means that if two solids specified the same input name you
could specify the input once in the config.Environment object
and then it would be filled in. However this also broke
modularity as collision could occur between solids that didn't
know anything about eachother.