Initial commit of the user-facing Dagster type system
This PR adds a user-facing type system to dagster. Instead of
"typing" solids by adding typechecks to input and output callback, one
can instead specify a type. This type itself knows how to do a
typecheck.

InputDefinition and OutputDefinitions now accept an optional type
argument (it defaults to the new Any type). This type
knows how to check to see if the runtime python type is compatible. It
also provides "default" sources and materializations.

Types now also accept nulls by default. We will opt into non-null
types just as GraphQL does.

Future enhancements could include a set of expectations along with
python type checks as well. This way expectations and the type system
would be unified.

There's a fair bit of noise in this PR at I used the addition
of the type system as an opportunity to clean up a bunch of stuff
to demonstrate value. The "pandas kernel" is much more simple, as
an example.