Execution Rewrite: The Compute Node
This is a substantial rewrite of the execution pipeline and will
open up a whole new range of explorations and more.

The core architectural change here is that there is an entire
new system of representation for the *execution* graph of computations
(compute nodes) rather than just the PipelineDefintion.

The compute node graph is constructed from the PipelineDefinition
and input config. A number of things happen during this construction:

1. All sources, materializations, expectations, and transforms are
modelled within the same abstraction: the compute node. All argument
checking and validiations are done during compute node graph
construction time, before any execution proceeds. This is similar to
validation as a separate step in the graphql universe.

2. Compute nodes can have multiple inputs, multiple outputs, and a
configuration dictionary (arg dict). This allows for maximum flexibility
and for it to be generic enough to handled the different concepts listed
in bullet 1. For example a source is just a compute node with *no*
inputs, a single output, and an arg dict. A materialization is a compute
node with an input (coming from the output of of a transform), no
outputs, and an arg dict. A transform can have multiple inputs and
(now) multiple outputs, but it current doesn't support an arbitrary arg
dict. Expectations are also a compute node.

3. This will allow for dynamically configured dependencies as opposed
to the more static construction. We are already using the dynamic
compute node graph to change based on policy. If, for example, you
do turn off expectations those compute nodes are never even
constructed. If they are they dynamically placed before the appopriate
input or after the appropriate output and set up to execute in parallel.

4. Sources and materializationsn as separate concepts are going to
largely dissappears. I am still toying with how to expose them to the
higher level config and pipeline definition API, but if they exist they
will be largely symbolic.

5. More sophisticated expectations execution will happen via this
mechanism. For example I fully anticipate that expectation compute nodes
will have two outputs, one for the value being tested and passed
through, but the other for the expectation result. This will allow the
framework (or the user in some capacity) to add compute nodes that act
on the expectation.

6. Composability: This is also an interesting angle for putting
composability into this execution engine. It should be totally possible
to build "composite" compute nodes that consist of a DAG of compute
nodes. When taken as a whole they will have a set of inputs, a set of
outputs, and a set of configuration.

7. Visualizations and debugging: This will be a powerful tool of
visualization and debugging. One will be able to set config, apply it to
a pipeline, and previz *exactly* what is going to execute in a very
granular way. Execution should be able to be reported through this
mechanism as well.

Next up I'm going to leverage this architectural change to change the
dependency API for pipeline definitions.