Move from "Argument Dictionary" to proper composited type system
I decided that this move should be made before 0.2.0. This is, IMO, a
much more sound system of managing config.

Instead of having "argument dictionaries" everywhere, configs take a top
level type to describe themselves. I've implemented one composite type,
the accepts a set of Fields and then produced a python dict for
implementors to process at runtime.

This system has a number of advantages:

1) Configs can be arbitrarily nested and grouped together. This already
would have been immediately useful.

2) We can move to strongly-typed APIs for configs. Instead of vending a
dict, vend a namedtuple, for example.

3) Even further that 2) this is a step towards a formalized "resource"
pattern for configs.  Instead of the type producing a dumb struct, you
could imagine the config producing a richer "resource" object that
manages connections or so forth. Or alternatively a system stacked on
top of this one that does that (and requires configs of a certain type).

4) This sets up the ability to support single value config quite easily.
I'm not sure if this is a good idea, but it is possible now if we allow
it.

One clear deficiency is a regression of error messaging. However, we
weren't testing error messages properly the entire time!
(see https://github.com/dagster-io/dagster/issues/135) So we need to
go through and audit type-system-related errors and figure out a way
to construct them in this new structure.