lambda_solid and solid
There are now two flavors of decorator.

1) lambda_solid. Takes inputs. No config or context. Produces a single output. Only
return naked value out of the transform.

2) solids. Takes inputs, produces outputs. Accepts context and config.
The transform can return a naked value, a Result, or a generator that
yields a sequence of Results.

This satisfies the need/desire for "simple" and introductory use cases
for solids (in the form of a lambda_solid) relatively elegantly but then
has a decorator-style solid that can cover nearly all use cases.