Run id and context reentrantcy
This is the beginning of a series of changes to allow a single
pipeline to be executed across multiple process. This PR adds a run_id
concept to the ExecutionContext as well as creates a new style of
"re-entrant" construction that will allow a user to seed a
ExecutionContext with a pre-existing context stack (that could include a
run id).

This will eventually be able to be driven by config, and we will be able
to restart execution of a subset of a pipeline with a prefilled context
stack.