Simplify implementation of dynamic loading, handle case where initial load fails
The problem with the current implementation of DynamicObject is that the tuple contains both the `module` reference and also the evaluated result. If you can’t evaluate the function, you can’t obtain a DynamicObject, which means you have no module handle to use to call `importlib.reload`.

To fix this I’ve split the loading into a two phase process—we create a `DynamicObject` with just the module and function name, and then evaluate the module separately. This allows 1) the evaluation to be done in the try catch and 2) means the initial load can fail and still have a DynamicObject with a module reference. I also made `eval` idempotent so it can also replace `reload_dynamic_object`.