Config type optional behavior (#248)
This PR pushes more of the behavior to determine defaults in the system config type system. This makes it so that one does not have to specify config values if all the values down in the tree are optional and defaults are filled in properly. This increases the usability of the config system.