Default value
We were accessing the default value without checking to see if it
actually was set in the graphql layer. This caused a hard failure.

Also adding better testing to dagit. I grabbed the graphql query as it
exists now from the chrome dev tools and plopped that in a unit test