Attribute types (#258)
* Add Type Attributes

Add the formalized notion of "type attributes" to the dagster type
system. The short term need is the need for the browser to distinguish
between builtin, system config, and normal types. However we will need
more of these attributes (having is_scalar and is_composite in here would
probably make sense, for example).

Added to both the internal type system and the graphql schema

* descriptions

* fix test

* final cleaup