Fix hot reloading not working due to new type coercion checks
Turns out two of the last edits to the hot-reloading PR broke it. This was actually super interesting to track down. When we allow the `reloader` module to reload the Dagster source, it causes `RepositoryDefinition`, etc to be reloaded, but only in the user’s pipeline code. Instances of these new copies of the classes do not pass instanceof checks against the old versions of the classes, causing the Dagster CLI (which is doing the hot reloading and was not reloaded) to think that you’re providing the wrong object types.