New error handling for config values (#283)
The primary goal of this PR to add an error detection pass to config parsing to allow the return of multiple errors in a given config document. This will enable much richer error reporting and a better config editing experiences.

Creating this PR also exposed some pretty serious cleavages in the type system, which this begins to address. There is still a lot of work to do.

Some highlights:

This PR separates the notion of evaluating a config value and producing a runtime representation from that config value from the notion of checking a type at runtime (between the output of solid and the input of another solid as an example). They are different operations with different semantics. Returning a stack of config elements as part of a runtime type check between two solids makes no sense, as an example.
Instead of hardcoding the weird behavior of contexts (only one field can be valid) I created the generic notion of a "selector" type, where the user can specify only one config element. I believe this will come up in other contexts. This is this type systems's equivalent of an input union type.
I could have kept going with this PR but I had to tie it off sometime. The next logical step after this is to great a GraphQL interface for all the relevant classes in evaluator.py.