Execution Lifecycle Change (#314)
This is a significant change to the ExecutionContext lifecycle and worth review and contemplation. Previously users directly created ExecutionContexts when they specified custom PipelineContextDefinitions. However, given that ExecutionContexts are immutable (outside of the context stack) this left the system with little flexible in terms of modifying the ExecutionContext or controlling its lifecycle.

This changes the contract of the PipelineContextDefinition so that the user instead returns an instance of ExecutionContextUserParams. This allows the user to specify some parameters to the context object, but the actual construct of the ExecutionContext object is in the system. This allows the system to interleave system-defined concepts along with user-defined concepts.

This change is used in this PR to add an extra ReentrantInfo parameter to execute_pipeline. This way the caller can provide a run id or context stack entries. We will leverage this in order to execute pipelines from within dagit, where we want to know the run id a priori, as well as inject a new system-specified logger.