Adding Execution Plan Step Logging (#318)
We now emit a well-formed event for every step of the execution plan. This is definitely exposing some warts of this subsystem.

It's probably too much work to add a typed event message.
The context stack makes thing difficult to reason about and test.
We need to add a property unique key to every step in the execution plan that is stable between process invocations.