More cleanly separate configuration and runtime type systems (#425)
The idea is to cleanly separate the notion of a types schematized config representation and its runtime structure. Before this was totally commingled, and there's really no reason for it to be. A config for a very complex runtime type could be just a string; and a config for a string could be a complicated config representation. There's no reason to couple the two.

So this extracts out the notion of Configurability into a set of mixins. This means that one can declare a type that is not configurable, or configurable by an arbitrary config schema, and that notion of the degrees of configuration is totally separated from the runtime type. I'm pleased the results of this so far.

The clarify results in better naming. For example instead of "is_list" the method is now "configurable_from_list" which is far more accurate.