Add flag to use synchronous execution manager for dagit (#427)
This is a stopgap to get a version of dagit which won't choke on Taylor's pipelines, while we debug the issues in the MultiprocessingExecutionManager. The (new?) issue here is that for some reason the SynchronousExecutionManager isn't actually streaming logs across the websocket -- they all come in at once when pipeline execution is complete.