Setup Improvements + Python 3.7 (#501)
This adds dev_env_setup.sh which should be a nice one stop shop for setting up a dev environment. I used this to setup up pyenv locally on my machine and to get python 3.7 up and running. Note I upgraded pyarrow requirement to ensure a version that is python 3.7 compatible.

Used my own docker account for the 3.7 image