Add escape hatch for uniqueness to ease upgrade pain (#493)
This is essentially a clarify-only feature. But we use clarify as a testbed for stress-testing dagit etc so I'm inclined to add this support, at least temporarily. Fixing clarify to have unique solid definition names will be a project given the complexity of the code and solid generation system. 

Made issue https://github.com/dagster-io/dagster/issues/523 since this is not obviously a win. However this PR unblocks viewing things in clarify repo for now.