Debounce per-subscription to avoid duplicates between initial fetch and events (#544)
This took a while to track down, but it turns out the duplicate logs were caused by a timing issue in the debounce implementation. The scenario is:

1) Log an event "A"
  + adds it to the _log_sequence
  + adds it to the DebouncingLogQueue

2) Request streaming logs via GraphQL
  + returns an initial payload immediately containing the current _log_sequence:  ["A"]
  + subscribes via `pipeline_run.subscribe` to receive debounced stream of new events. 

3) debounce queue timout expires, flushing the queued contents. "A" is emitted again.

I think the fix for this is to implement the debouncing per-subscription, so the debouncing begins at the start of the subscription and never contains previous events. This might be bad if we had many subscriptions going since we're switching from 1 queue to N queues, but in this case I don't think it matters.