Initial support for scaffolding config in the pipeline execution tab (#552)
This PR makes it so that if you 1) open the execution tab for the first time with no saved pipelines or 2) click "Add..." and create a new config yaml, the content of the document defaults to a nice scaffold.

The scaffold includes a header note and definitions for all required fields that are isOptional:false in the field tree. Right now, I'm inserting some bogus default values so it's easy to see what type of value you need to provide (For example, path: "/file/path/here"), but in the future I'd really like to make these higlighted / templatized in some way.