Add dagster_type decorator (#576)
This adds two methods of creating dagster_types. One is the dagster_type decorator. This is used for new classes. The other is "make_dagster_type" which can help out with existing classes that one cannot modify. I decided to go with a monkeypatching approach. The other option would be to have a cache of classes for the whole program. I thought this was the better option all things consider because I like to avoid these type of global caches when possible.