In dagit config scaffolding comment out all but one "option" in a selector (#584)
This PR exposes the isSelector field to GraphQL and uses it to improve the scaffolding. Only one option is enabled and the others are commented out.

I went through a bit of extra work here to make the comment style match exactly what CodeMirror does on it's own. We generate a JS object tree with keys like COMMENTED_(DEPTH)_fieldname, convert the object to a YAML string, and then replace the COMMENTED_ prefix with a # character and (DEPTH) number of trailing whitespaces once the YAML has been pretty-printed.