Dependency Management in Runtime/Config Type Schism (#610)
Continuing to work out circular dependency issues cropping up in the runtime config system. In this case, we very cleanly separate the user-facing APIs for Field where one can pass in the enum representing a built-in or a config type or a runtime type from the field internals. The built-in schemas of ints etc must only use the internals version of this, otherwise we get into a circular dependency hell.