Enum bones (#684)
This is the beginning of adding enums to the dagster config type system.

This just tests basic functionality and ensures the basic error checking works.

Some things of note:

1. The public API is `Enum` which name collides with the python `from enum import Enum` which could be annoying.
2. EnumValue represent each enum value. I have opted for "config_value" and "python_value" to indicate a) what would represent the enum in a config doc and b) how that is represented in memory respectively. In GraphQL we called this 'name' and 'value' which I found endlessly confusing. This is explicit as the expense of verbosity.

Next steps:

1. Add to the GraphQL Config API.
2. Add to the Config Editor.
3. Add a decorator-based API where one can directly annotated native Python Enums. (this would also solve the name collision issue as we could have a @dagster_enum decorator or something).

I added the bit as that does coercion into the underlying python representation of the enum in the "default applier" so that isn't the best name anymore. Need to think of new name.