Implement Marshal and Unmarshal as execution steps (#765)
It should have always been like this. Instead of doing unmarshalling and marshalling outside of plan execution, do it within it by injected a step per input and output that needs to be marshalled. This unifies error handling and partial execution, and will assist as we build debugging features and viz on top of the execution plan abstraction.