Info --> Context (#798)
This diff changes the top-level object from info to context everywhere. What was RuntimeExecutionContext is now LegacyRuntimeExecutionContext. (With the new API in place we can do subsequent refactors to eliminate this object). All top-level fields that were available on the legacy context object have been hoisted to the formerly-info-but-now-context object. TransformExecutionInfo is now TransformExecutionContext.

This also renames the `config` field on the formerly info object to `solid_config`. Using `context.config` to access the config for the solid was super confusing, so i changed it to `context.solid_config` which is much clearer.

This is a non-breaking change (for now). Old code with `info.context` syntax will continue to work as will `info.config`. At some point we should deprecated and eliminate, but I think that is an 0.4.0 conversation.

Follow on work:

- Eliminate LegacyRuntimeExecutionContext
- Rename/refactor info object passed to context and resource creation. (Maybe call it InitializationContext?)
- Rename/refactor info object passed to expectations functions.