Do not recursively watch all files in parent dir
If I create a directory in my home dir:

`mkdir ~/test`

and then run dagit from inside that directory:

`cd ~/test && dagit`

the current code recursively watches all files in the parent directory,
which in this case is home. Not only does this seem incorrect, it also
makes dagit load very slowly on my machine because there is a ton of
stuff in my home directory. The current working directory is already a
dir, so this patch removes the (accidental?) dirname call.