Serializable execution plan (#866)
This creates a new API to execute plans in such a way that guarantees that it is serializable. We then use this API to implement the graphql backend in order to test it and consolidate code paths.

The general process for using this API is to construct the execution plan in the parent process before execution. This is fine because the parent process orchestrating the execution of subplans will have to create the plan to determine what subset to execute.

This effectively serves as a forcing function to make step events serializable.

Next steps:

1) Use this API to implement the multiprocessing executor.
2) Use this API within the dagit pipeline execution manager. 
3) As part of 1 and 2, expand the number of events that we are streaming through the step events channel. Part of this work will be to actually merge the notion of ExecutionStepEvents in the core executor and the "EventRecord" events in dagster.core.events. We have two parallel event streams which is wasteful and annoying.