Intermediate Value Manager (#870)
This adds the notion of an IntermediatesManager to manage intermediate values. This is definitely a "midpoint" diff as there are a lot of follow up tasks. However, I wanted to get to a checkpoint, discuss direction, and then get this back into master. There are also a few refactoring changes that come along with this PR.

The primary value add of this PR is that the multiprocess engine no longer needs to hold all the input s and outputs for the pipeline in memory. Instead, if that caller needs to access these values they are accessed by demand, through the manager. Concretely, this means that produced values are accessed via the file system in the multiprocess case.


A few refactoring things:

- StepOutputHandle now only contains the step_key rather than the entire step. This means that it is directly serializable across processes. (It also eliminates all the custom hashing stuff implemented on that class in order to make it useable as a dictionary key). It should have always been like that. The only downside is that we need to do a little more work in the GraphQL engine to thread the execution_plan down to places where we want the entire Step (e.g. dependsOn on StepInput)

- success_data has been renamed to step_output_data. We will be adding more event types to this stream and success_data will soon become ambiguous. This rename should have been done when the class was renamed anyways.

- StepOutputHandle as used in the execution creation pathway to pass around full steps. This has been replaced by the SingleOutputStepCreationData tuple.

Next tasks:
1) Use the "serializable" execution path in the multiprocessing. This will replace the BS
custom serialization that is in the multiprocessing engine
2) Marshal higher quality error information across serialization boundaries.
3) Use the persistence and serialization infrastructure in the intermediates manager
4) Come up with common, rationalized management of temp files. This will likely be a first class resource similar to context.log. context.files or something.

I believe these tasks actually overlap with both @puma314 and @mgasner 's tasks. I think proceed in parallel and then consolidating later is a fine strategy, but we should be aware of what eachother are doing.