Separate User and System Context objects (#879)
In anticipation of the file resource work, I just separated out a layer of context objects that are user-facing versus system-facing. The logic is that we will have a place in the file system/whatever store for storing marshalled intermediates that we don't really want the average user to interact with.