Rip out ExecutionStep injection; Make features marshalling specific
This PR removes the ability to do step injection generically. This
feature was too complicated and too general, and only implemented to
enabling marshalling. Instead we implement marshalling directly in
the execute_marshalling function. We keep the graphql
interface to subplan execution effectively constant (except for
a couple error conditions) while changing the internals.

The next step after this diff is will be instead to enable marshalling
directly via the intermediates manager. A remote subplan execution
will work by setting up the intermediates *prior* to the execution
of that subplan.