Extracting out child process executor (#893)
Extracting out child process executor for the multiprocessing engine. This is starting to overlap with the pipeline execution manager more and more, so merging this should be a goal. In particular the pipeline execution manager handles crashing processes and multiple child processes, so this is strictly inferior in terms of features.