Make run storage configurable to inmem versus file-based (#937)
Introduces the notion of a config-driven decision whether or not to use the filesystem or inmemory for storing runs. The CLI defaults to filesystem; the API called from python defaults to inmemory (for testing). 

This setups the ability to also have an S3-backed intermediates manager and run storage concept.

Next steps on my end:
   - Using this mechanism to replace the execute_marshalling codepath (instead the values will be source from the intermediates manager).
   - Instead of hardcoding to pickling, using the serialization mechanism attached to the type system. Will be needed for spark integration.