Improve error message when an exception leaks from execute_pipeline when using dagit (#1054)
This is a short term mitigation to a deeper issue.

This was exposed when an error leaked out of our intermediates manager when trying to serialize and deserialize the spark dataframe being passed between solids. @mgasner is aware of the issue and is working on a fix.

A few things to note here:

When an error occurs in the intermediates manager such as this, this should probably be noted as a step-level error rather than a pipeline-wide level error.

However, even with this fix, there are going to be errors that cause pipeline-wide errors that are not attributable to a particular step. We should emit a dagster event for this case and render it in some way in the execution pane.

We should also make it so that when can get to the stack trace model popup from within the log viewer, not just the execution plan pane. This could be a nice feature to implement at the end of @alangenfeld's workstream to unify our event systems. It would require that the log viewer have access to the array of strings that comprise a proper stack trace, rather than just smashing it into a string, as this PR does.

Our error message handling is a bit of a mess, as noted in pipeline_execution_manager.py:38

        # Currently it is the user_message that is displayed to the user client side
        # in dagit even though that was not the original intent. The original
        # intent was that the user_message was the message generated by user code
        # communicated directly to the client. We need to rationalize the treatment
        # of these different error messages