Basic re-execution via a new "play" button beside finished / failed steps (#1208)
This PR adds a small play button that appears when you hover over a step that has succeeded or failed:

<img width="345" alt="image" src="https://user-images.githubusercontent.com/1037212/55909969-19efa780-5b92-11e9-8c2d-2c2cae8fd340.png">

When you click it, it starts a new execution of the pipeline with the same config and all of the outputs of steps used by the selected step captured in the `reexecutionConfig`.

In order to make this work on the demo pipeline, I needed to specify persistent run storage in my pipeline config explicitly. Otherwise it says it can't find the previous run once it's started:

```
storage:
  filesystem:
    base_dir: "~/Desktop"
```

The new run opens in a new tab similar to the way fresh runs work.