Add computed execution_plan metadata
Summary:
This is the beginning of allowing users to provide metadata to
generated for storage inside of execution steps. The user can provide
a function that generates a dictionary given the passed in config. This
for example could be a filled out spark-submit command, or a rendered
sql template, as examples.

Test Plan: Included unit test

Reviewers: alangenfeld

Differential Revision: https://dagster.phacility.com/D6