Execution step metadata (#1305)
This is the beginning of allowing users to provide metadata to
generated for storage inside of execution steps. The user can provide
a function that generates a dictionary given the passed in config. This
for example could be a filled out spark-submit command, or a rendered
sql template, as examples.

To disambiguate from metadata, indicate that tags are used specifically for logging