Flow Materialization all the to the GraphQL layer
Summary:
This is the start of the restructuring alex and I discussed this morning

The overall plan is to introduce a Materialization type in the GraphQL domain
and expose that same type on both StepOutputEvent and StepMaterializationEvent. This
way we can expose this data to UI identically and have a single, coherent way of
handing the way that we materialize external artifacts. The first step is capture
the Materialization event and flow it all the way up to the GraphQL layer.

Steps after that:

  - Also change step output to contain a Materialization object. Flow up to graphql layer
  - Change Materialization to have proposed structure (desc, path, url, metadata)
  - Change GraphQL layer to have a Materialization type and attach that to corresponding
    graphql types and update UI

Test Plan: Unit tests

Reviewers: alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D13