Extract out s3 object store into dagster_aws
Summary:
Extracting S3ObjectStore into dagster. This would be step one in making the storage system truly pluggable. The next step would be to have it so that we can instantiate a pipeline or a repository and passing in an object that says "hey this additional storage strategy is available"

The diff also demonstrates that it is pretty brutal to change a dep like this. You have to touch dockerfiles, build.sh's, tox.ini's and setup.py's all over the codebase.

Test Plan: Included tests

Reviewers: alangenfeld, max

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D29