Create EnvironmentSchema
Summary:
This is the first step in making an environment mode
that one passes in to execute_pipeline instead of using a
context selector. We must extract out configuration schema creation
out of the pipeline definition. An EnvironmentSchema will now be a
function of pipeline + mode

Test Plan: Unit tests

Reviewers: max, alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D57