Add mode to graphql graphql inteface updates
Summary:
This adds mode to the graphql interface only. This adds a weird in-between
state where we have to provide a default mode sometimes and I've had to added some temporary
weird code where we just arbitrarily pick a mode. This is tracked by
https://github.com/dagster-io/dagster/issues/1343. Otherwise this is mostly done with
unit tests, both the execution code path as well as the read code path. Also made the full
production query (or an approximation of it at this point) a snapshot test in order to
exercise the read path.

The next step is to add the mode stuff on the UI side.

Test Plan: Run mode and modeless pipelines in dagit

Reviewers: max, alangenfeld

Reviewed By: max

Differential Revision: https://dagster.phacility.com/D70