Nuke the monads from orbit
Summary:
This is a big refactor of the graphql codepaths to eliminate the "Either" monad-esque
and instead just use vanilla functions for composition. In order to communicate error conditions
we just throw a new exception type, UserFacingGraphQLError, which wraps the DauphinError. We
then add a error boundary/decorator @capture_dauphin_error to catch this and shuffle the error to GraphQL.
This is sort of using doing exception as control flow (pythonic!!!) but it only signals
the end of the computation, rather than any sort of resumption semantics, so I think it is fine.

The end result is certainly much much easier to deal wiht.

Test Plan: unit tests

Reviewers: max, alangenfeld, natekupp

Reviewed By: natekupp

Differential Revision: https://dagster.phacility.com/D189