Add support for composite solids to the Explore tab
Summary:
This diff adds breadcrumbs to the explore tab and allows you to double-click into composite solids, which have a new purplish + double bordered appearance.  When you're within a composite solid, the DAG shows a containing box indicating the solid you're within, and when you don't have a selection you can view the parent solid and its input/output mappings in the right sidebar.

Hovering over an input / output of a composite that is connected to one of it's child solids displays the name of the child. We should be able to do this the other way around as well. Will investigate that in a stacked diff.

This diff also changes the URL structure within the explore tab. The path segments are the handle IDs, and a trailing slash indicates you're within the last path component rather than just selecting it. It seems that the solid HandleIDs are generated as dot-separated paths, so there's a bit of duplication in the URL currently, eg:

```http://localhost:3000/composition/explore/add_four/add_four.adder_1/add_four.adder_1.adder_1```

I think we could fix this by allowing the client to rely on this HandleID structure, but I wasn't sure if this was a good idea.

(Note the purple arrows were removed!)

{F17086}

Test Plan: Updated the snapshot tests

Reviewers: alangenfeld, schrockn

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D205