Do not pin to python37 executable in append_generated
Summary: Hardcoding to python37 executable in this script creates dependencies on the local setup of your development environment. For example I manage python dependencies via pyenv and this command is not available. If we need to pin to specific versions with should check that within append_generated itself. Or we should figure out an alternative approach for hiding autogened ts files

Test Plan: Run generate types

Reviewers: alangenfeld, bengotow, max

Reviewed By: bengotow

Differential Revision: https://dagster.phacility.com/D301