Improve times for airline-demo tests
Summary:
This is a few changes which bring airline-demo tests down to
5-6 minutes reliably. The biggest variable is whether or not the
integration image is on the machine or not.

A few changes here:

1) Use the integration image for the airline demo

2) Add the triggering of the maven d/l for the airline demo currently
in the airline demo docker image to the integration image. Also install
libqpq-dev

3) Dramatically simplify airline demo Dockerfile by using integration
image and then consolidating to a single ADD command and a single
pip install. Consolidating these commands saves a lot of time,
interestingly.

I tacked on a "-v2" to the docker image tag to make cutover
atomic.

An optimization we could do is preinstall all the integration images
in a custom AMI. When the image is not there it adds about a minute
to the test run.

It actually appears that the events-demo is now the long pole.

Test Plan: Buildkite

Reviewers: natekupp, alangenfeld, max

Reviewed By: natekupp

Differential Revision: https://dagster.phacility.com/D326