(storage-plugin-system-4) Eliminate RunStorage and new strategy for type plugins
Summary:
This is a change to the type storage system plugins. We have
added a function to the TypeStoragePlugin interface called
applies_to_storage that determines whether or not to use a
plugin based on the SystemStorageDefinition that is active
for the current execution.

I have also renamed the plugins passed to types "auto_plugins". These
are plugins that the system attempts to load whenver that type is
included in a pipeline.

Extensions of this would include:

1) Pipeline-specific overrides. Imagine if a user wants to either a)
   override the type storage plugin. (E.g. they want to persist dataframes
   in s3 in a specific way but still use the out-of-the-box system storage
   definition) or they want to have a type storage plugin for their own
   system storage definition.

2) Config-driven type plugins. Allow users to configure type plugins.
   For example maybe they want to persist spark data frames as csv
   instead of parquet.

Also deleted the unused doc snapshots.

Test Plan: buildkite

Reviewers: natekupp, max, alangenfeld

Reviewed By: natekupp, max

Differential Revision: https://dagster.phacility.com/D341