Rework dagstermill context
Summary:
This allows users a full-fledged interactive context using dagstermill.get_context(), including the
ability to shim in input and output definitions, config, and resources, or to load an existing
pipeline or solid directly (using an in-memory Python API), or using handle CLI args. It
substantially simplifies the in-pipeline (dagstermill-managed) context and serialization handling
by requiring that dagstermill solids execute only when a pipeline handle is available.

Test Plan: Unit

Reviewers: schrockn, natekupp, alangenfeld

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D446