Tighten up input behavior
Summary:
This tightens up the behavior of the step input event and tests
that behavior by running pipelines without raise_on_error. This makes it
so that the STEP_INPUT event is emitted *any* case where the type check
fails, rather than the case where it emits an explicit Failure
exception. That means a usage error or an exception of an arbitrary type
also emits this error type. Essentially the only advantage of the
Failure exception type is the ability to attach metadata

Wanted to get right before changing STEP_OUTPUT to match behavior

Test Plan: Buildkite

Reviewers: alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D489