[dagster] fix multidep of one issue
Summary: Previously we were holding all deps as a single array and differentiating by if there was just a single entry. While this might be reasonable if you were always statically encoding dep graphs and only used multi when N was > 1, it falls down when ever you are programmatically generating over a varying input size that ever might be 1.

Test Plan: added test

Reviewers: schrockn, max, natekupp, sashank

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D563