Fix pandas perf issues -- Remove valueRepr from the system
Summary:
Jack reported in #1553 that overhead in dagster was
surprisingly high. This appears to be mostly the result of stringifying
outputs as part of logging during the execution process. Unfortunately
for data frames this appears to do something very expensive, like touch
everything in the data frame or similar.

Given that there are no guarantees around repr() and str() in python,
this diff simply removes the entire thing from the system. The core
will no longer get the python representation of objects for logging
nor will be stream that to the client.

Test Plan:
Buildkite. Run case provided in https://github.com/dagster-io/dagster/issues/1553
under python cprofiler. Confirm that overhead is no longer there.

Reviewers: #ft, alangenfeld

Reviewed By: #ft, alangenfeld

Subscribers: alangenfeld

Differential Revision: https://dagster.phacility.com/D625