Consolidate Airflow and Dask GraphQL client logic
Summary:
This diff cleans up the divergent code paths that `dagster-airflow` and `dagster-dask` used to execute pipelines through `dagster-graphql`.

Now, a single function is provided in `dagster-graphql`, `execute_start_pipeline_execution_query()` to execute the pipelines and return a list of DagsterEvents.

This is in preparation for a following diff which will consume the dagster events to inform Airflow that a particular task was skipped using the Airflow `SkipMixin`.

Test Plan: unit

Reviewers: #ft, max

Reviewed By: #ft, max

Subscribers: max

Differential Revision: https://dagster.phacility.com/D677