Avoid repeated calls to readFragment to shorten new-logs critical path by 30%
Summary: It turns out we spend a rediculous amount of time reading the data for the pipeline
and then writing it back to the Apollo cache because their data store is heavily normalized.

I think that long term we should consider streaming the logs via a raw GraphQL fetch and not
try to store the logs into Apollo-client at all. In the very short term, we can switch from
read+write, read+write, read+write to read+write, write, write, write. Because we know that
the run data will not be updated from elsewhere.

Test Plan: Run tests

Reviewers: #ft

Subscribers: