Overhauled run history view
Summary:
This diff replaces the old run history view which you could view for a particular pipeline with a new one that is at the top level of the Dagit UI and allows you to explore the config, etc. of previous runs and view summary stats.

This diff also allows you to pass ?q= to a run URL to pre-filter the logs, and pass ?config=&mode=&solidSubset=[] to the /:pipeline/execute URL to create a new tab with that configuration.

Right now this view auto-refreshes every 15 seconds and requires the logs for every run (to compute summary statistics like the number of failed expectations). This will be fixed later.

Test Plan: Run snapshot tests

Reviewers: #ft

Differential Revision: https://dagster.phacility.com/D858