Remove lambda_solid from tutorial and examples
Summary:
This has always been a bit problematic and now we have evidence that it
is actively causing confusion for folks getting up to speed. Post py3 type
annotations this is actually more reasonable to deal with, and it only
requires having a placeholder variable for the context variable.

Test Plan: Look at docs

Reviewers: #ft, alangenfeld

Reviewed By: #ft, alangenfeld

Subscribers: prha, alangenfeld

Differential Revision: https://dagster.phacility.com/D874