Use dagster-graphql for mock response generation
Summary:

This diff replaces the old system that generated mocks for the snapshots tests as well as my `curl` mocks with dagster-graphql. Previously, you'd run dagit with a special environment variable, and it'd open a web page with JSON showing mock requests / responses and you'd paste it into mockData.tsx.

The new solution is a simple script that pulls the mocks each test needs, fetches them, and writes them to a file.

The only slightly odd thing about this approach is that we're using `jest` (the test running framework) to run the scripts. This allows us to leverage a ton of magic within Jest that both compiles typescript on the fly, and simulates a DOM environment within nodejs. This is why we can import the components and their queries, and also import `fs`. Technically we could set this up on our own for these scripts, but it seems like maintaining three separate targets (web, jest, and scripting) would be a pain.

Test Plan: Run shiny tests

Reviewers: #ft

Subscribers: