force run view under pipeline in url namespace
Summary:
The Run history refactor split out the runs view to be independent of a given pipeline.  The TopNav component assumes that there is a selected pipeline, which gives the context for the `Explore` and `Execute` tabs.

This diff changes a single run view to still be scoped under a specific pipeline, changing the URL to include the loaded pipeline name.

Any loads of a single run view (`/run/:runId`) will redirect to a URL containing the pipeline name (`/:pipelineName/run/:runId`)

https://github.com/dagster-io/dagster/issues/1701

Test Plan: clicked to runs view, clicked through to single run view, saw appropriate url with pipeline name, and top nav dropdown populated

Reviewers: #ft, bengotow, sashank

Reviewed By: #ft, bengotow, sashank

Subscribers: sashank

Differential Revision: https://dagster.phacility.com/D909