[RFC] DagsterInstance
Summary:
`DagsterInstance` represents the set of Dagster systems that are interacted with when executing a pipeline but may exist beyond the duration of an execution. The initial motivating systems that we have today are `run_storage` and `event_storage`.

If you want to execute a pipeline and persist no history of it such as for test purposes - `DagsterInstance.ephemeral()` is used. This persists nothing about runs to disk and uses a temporary directory for intermediates/file_manager storage.

If you want to execute a pipeline that uses the files system to persist information across runs such as for re-execution, but once that test is over you want everything cleaned up, `DagsterInstance.local_temp()` is used.

Tools like `dagit` and `dagster-graphql` on the other hand will use `DagsterInstance.get()` which in the common case uses `$DAGSTER_HOME` and `$DAGSTER_HOME/dagster.cfg` to determine the configuration for the `DagsterInstance` to use. This gives you a consistent local experience across uses of the tools.

For easier onboarding - when a someone uses dagit for the first time we don't force them to set $DAGSTER_HOME and instead hold open a temp directory in the "watchdog" `dagit.py` parent process that can stay open across restarts allowing someone to maintain a temporary history of runs while debugging an issue which is all cleaned up when the dagit parent watchdog process exits.

In this diff `DagsterInstance` starts alleviating a lot of confusing data flow and cross object communication by giving us a place to start flowing data in one direction about runs and events occurring in a particular `DagsterInstance`. This is key to accelerating our construction of operational tools that allow users to monitor these events. This set up gives us a pretty clear place to slot in "remote" instance types which will be a common setup for "production" deployments of dagster.

Test Plan: unit tests

Reviewers: #ft, schrockn

Reviewed By: #ft, schrockn

Subscribers: max, schrockn

Differential Revision: https://dagster.phacility.com/D936