Add lexer/parser for subselection language
Summary:
I had a bit of time and am holding off on pressing forward on the runs db pending the dagster instance refactor.

So I made a first stab at a lexer/parser for the selection syntax to select many solids. This is largely based on
the dbt syntax. This is a bit overkill given the current language, but I strongly suspect we will make this
more complicated as time goes on, so I set it up properly.

Tracking progress in https://github.com/dagster-io/dagster/issues/1552

What this will enable (eventually) is typing at the command line or in
dagit UI. "+some_solid" which would select all solids necessary to
execute up to some_solid.

Test Plan: Buildkite

Reviewers: #ft, natekupp

Reviewed By: #ft, natekupp

Subscribers: natekupp

Differential Revision: https://dagster.phacility.com/D939