Mock more of the GraphQL queries, add snapshot test for new runs page
Summary: This diff adds a snapshot test for the new runs page. In order to show an
actual run on the page, the script to generate mock data now kicks off a log_spew
before using dagster-graphql to query data.

I'm not super happy with how this snapshot testing stuff is working - I tried for
a while to add a test of the run details page, but generation of the mock data is
complicated because the run gets a random ID and all the timestamps cause the
snapshot to break whenever the mock data is rebuilt. Apollo was also throwing
cryptic errors I think are related to mocking the streaming logs query.

Long story short I think we'll need to move to slightly more granular tests for
the run details page. Will take another pass at this soon.

Test Plan: Run new tests

Reviewers: #ft

Subscribers: