Update S3 scratch bucket to fix tests
Summary:
Using a new scratch bucket on S3.

I had previously been using "dagster-scratch". I attempted to mutate the AWS provider of that bucket in Pulumi today, and Pulumi tried to delete and recreate it—but it can't do that, since the bucket has the same name, and recreating a bucket right after deletion is not supported (takes somewhere between 15 mins to multiple hours before it is available again).

This uses a Pulumi-suffixed bucket, which ensures that issue won't happen again—but we'll still need to find a way to keep dagster and Pulumi synchronized if we ever change the bucket name again.

Test Plan: unit

Reviewers: schrockn

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D978