Basic implementation of search through run history
Summary:

This diff adds a basic search bar to the runs list that searches pipeline name, mode,
and stepsToExecute (if present.) I'm planning to add autocompletion / tokenization in
a stacked diff.

This also includes a small GraphQL interface change - the type of stepKeysToExecute
was `(string | null)[] | null` which isn't quite right. If it's present, the elements
are non-null strings. Updated the dauphin representation and checked that it's now
(string | null)[]

Test Plan: Run tests

Reviewers: #ft

Subscribers: