[1/3] Display latest runs in schedule UI
Summary:
Display a list of runs and their statuses that were created by each schedule in the Schedule UI, as well as the date of the last run. You can click on the status dot to navigate to the log viewer for the run. The list refreshes every 15 seconds, just like the runs history page.

{F39971}

Depends on: D980

Test Plan:
```
cd projects/dagster/examples/dagster_examples/experimental
dagit -p 3333
```

Reviewers: #ft, alangenfeld

Reviewed By: #ft, alangenfeld

Subscribers: alangenfeld

Differential Revision: https://dagster.phacility.com/D999