(2/n) Create ScheduleStorage and refactor Scheduler
Summary:
Introducing a `ScheduleStorage` interface and `FilesystemScheduleStorage` implementation which manages reading and writing metadata files for `Schedules`. We pass only a `ScheduleStorage` to a `Scheduler`, which has two benefits:

- `Scheduler`s no longer need to store and manage `ScheduleDefinition`s
- `Scheduler`s don't need to manage metadata files

This simplifies the `Scheduler` interface, and the logic for consolidating `ScheduleDefinition`s and previous `Schedule`s is moved to the `ScheduleHandler` (previously called `ScheduleBuilder`.

Test Plan: unit

Reviewers: #ft, schrockn

Reviewed By: #ft, schrockn

Subscribers: schrockn, alangenfeld

Differential Revision: https://dagster.phacility.com/D1106