Move log stream out of Apollo, memoize row contents and fix 2x cascading renders
Summary: This diff makes the run log viewer fast.

It turns out there were several problems that had built up in here. We've been mostly implementing the "streaming GraphQL subscription which appends to an array" ourselves, but on each inbound update we update the data and commit it to Apollo's normalized store. We have no interest in a normalized store for log entries and Alex noted earlier today it's super slow.

This diff merges the RunSubscriptionProvider and LogFilterProvider into a single LogProvider that opens a websocket, sends the GraphQL request, and then accumulates the logs into our own array which is vended to it's children. The logs are not requested through Apollo at all. It turns out it's faster to retrieve the "initial" / old logs this way and not use the cursor at all, which is also interesting.

This diff also avoids throwing away the virtualized grid's cell size cache every time new logs arrive. We were doing this because the search / filters in the view hide items and invalidate the row heights. Now we generate a key from the filters - as long as the filters remain the same and we know the logs are append-only, it's safe to keep using the cached row heights.

It turns out we were also spending a good deal of time just rendering the actual cells. Because each cell is rendered and then re-rendered with it's fixed size, memoizing the cells /inside/ of the main row container is much better than memoizing them at the top level (where the `style` prop changes.)

Anyway. Here we go:

Test Plan: Run log_spew (faster)

Reviewers: #ft

Subscribers: