Fix GraphQL error when switching between pipelines with different modes
Summary: The error toasts that appeared when you switched pipelines in the execute tab were caused by the local storage provider, which was vending data for the wrong pipeline for one render cycle because it was keeping it's own state in sync with a useEffect (componentDidUpdate) hook.

Test Plan: Run and see no errors!

Reviewers: #ft

Differential Revision: https://dagster.phacility.com/D1099