(add-cancellation-7) Implementation of get_runs_by_status
Summary:
This will be necessary to reconcile and cleanup started runs when the dagit webserver restarts. On bootup it will find
started runs that are started but don't have an associated process.

Also changed the storage of enums to be enum_value.value as opposed to
str(enum_value). str(enum_value) includes the class name which actually
could threaten the 63 char limit. Also the behavior is more well-speced.

Also did some refactoring of run stores

Test Plan: BK

Reviewers: prha, #ft

Reviewed By: prha, #ft

Subscribers: alangenfeld

Differential Revision: https://dagster.phacility.com/D1110