Expose run filtering via GraphQL and implement pagination in Dagit
Summary:
This diff implements server-side run filtering and client-side pagination through runs in Dagit. There are two new GraphQL endpoints that map on to the run storage interface:

`pipelineRunsOrError` replaces `pipelineRuns`. Params are cursor, limit, and a selector with one of the following:

```
type PipelineRunsFilter {
  runId: String
  pipeline: String
  tagKey: String
  tagValue: String
  status: PipelineRunStatus
}
```

Example usage:

```
query {
  pipelineRunsOrError(
    limit: 1,
    cursor: "9e51122f-9c6f-47c0-b92f-4bd156f4f96b",
    filter: {
      pipeline: "log_spew"
    }
  ) {
    ... on PipelineRuns {
      results {
        pipeline {
          name
        }
        runId
      }
    }
    ... on InvalidPipelineRunsFilterError {
      message
    }
  }
}
```

I also added `pipelineRunTags`, which returns all of the key-value tags currently in use and implemented it in run_storage. This allows for search autocompletion based on tags.

```
query {
  pipelineRunTags {
    key
    values
  }
}
```

The UI has been made a bit more minimal - the filter box looks the same but only allows you to put in one filter, and the solid subset option is gone. Status was merged in since you can only filter by status if you do not choose another filter, and I removed the sort box. I also removed the "most recent run" box because it was really just a normal row pulled out of the table. (Originally it was going to show extra info.)

Pagination works by requesting (displayed + 1) rows. If we want to show 50 but find that there are 51, we know there's a "next page". A next page button appears at the bottom and clicking it advances the cursor. Because we use runIds as cursor and they are non-sequential there's no way to know what the cursor for "N-50" would be, so the cursor is stored in the browser URL bar and your browser back button will take you to the previous page. The filter / search is synced to the URL bar as well so you can share links to specific filters.

I also made the UI show the tags on the runs in the `Execution Params` columns. Boring styling for now.
(EDIT: This seems to not work - you can request `pipelineRuns { tags { key value }}` via GraphQL but the tags are never returned.

Test Plan: Run tests

Reviewers: #ft, alangenfeld

Reviewed By: #ft, alangenfeld

Subscribers: schrockn, prha

Differential Revision: https://dagster.phacility.com/D1150