Fix schedule --start-all command
Summary:
Previously, `all_schedules` was a property on `Scheduler`, but it was changed to an abstract method. The reference to `all_schedules` was not changed from a property access to method call in the code path for the `schedule --start-all` cli command.

This diff fixes that and adds tests for the `--start-all` command

Test Plan: `test_schedules_start_all`

Reviewers: #ft, natekupp

Reviewed By: #ft, natekupp

Subscribers: schrockn

Differential Revision: https://dagster.phacility.com/D1226