Make construction of ConfigurableClass-derived classes less error prone
Summary:
Per https://github.com/dagster-io/dagster/issues/1804, it is very
easy to screw this up as currently structured because of the order of __init__
calls in Python when multiple inheritance is in play.

Instead this pushes management of the inst_data property down to each concrete subclass, eliminating
the need for the __init__ in the ConfigurableClass class. This causes a little bit of code duplication
in exchance for no longer having an unexpressed, implicit contract for the __init__ call signature
and no need to manage MI __init__ shenanigans.

Test Plan: BK

Reviewers: #ft, max, alangenfeld

Reviewed By: #ft, alangenfeld

Differential Revision: https://dagster.phacility.com/D1265