Update get_run_tags to return unique tag values
Summary:
`get_run_tags` is a method on `RunStorage` which returns a list of `Tuple[string, List[string]]` where the first element of the tuple is the tag key, and the second element is a list of all the values for the given key. Previously, the values contained duplicates if two runs had a matching tag.

This caused an issue in dagit where the same tag value was being suggested multiple times in the filter bar: https://github.com/dagster-io/dagster/issues/1839

This diff updates the method to return unique values.

Test Plan: Load dagit, confirm tag suggestions are unique

Reviewers: #ft, prha

Reviewed By: #ft, prha

Subscribers: prha

Differential Revision: https://dagster.phacility.com/D1277