Repository-level solids explorer + graphQL root
Summary:
This diff adds the  graphql interface proposed by Alex in his feedback on D365. A new root query allows you to fetch solids and their invocations across the whole repo.

I think that having more repo-level concepts / pages in Dagit will inform how we want the new nav to work. This diff adds a solid explorer tab alongside Schedule and Runs. This view allows you to view all of the solids and see the same definition details visible when you click them in the Pipeline Explore tab. (I refactored a bunch of stuff to make our components more explicit about whether they require a definition, an invocation, or both, allowing the SVG and sidebar stuff to be reused.)

So far the coolest part of this UI is that you can filter solids by pipeline OR by data types. So I can click a solid with a "DataFrame" output and see all the solids that have a "DataFrame" input.

I think the actual design could use a bit of iteration, but curious to get thoughts on this v1!

Test Plan: Will add a snapshot

Reviewers: #ft

Differential Revision: https://dagster.phacility.com/D1323