Merge test and code files in tutorial
Summary:
In order to completely avoid the thorn issues
of importing things in and outside of packages, we instead
are just going to have the tests in the same file as
the example code. This allows the code to be included directly
verbatim from the file without dealing with the fact that in
our tests, we are executing in packages, whereas the user
will probably not be doing that in the tutorial context.

Test Plan: Run pytest manually against hello_cereal and complex_pipeline

Reviewers: max

Reviewed By: max

Differential Revision: https://dagster.phacility.com/D1408