[instance] remove features system
Summary:
This is currently unused and not the right implementation given newer abstractions like `ConfigurableClass`. I think we can do a better job when we need it (or just revert chunks of this diff).

My motivation for cleaning this up is to provide a cleaner "override" ability for loading an instance which I think we'll need for the execution server / run queue

Test Plan: bk & lint

Reviewers: prha, sashank, max

Reviewed By: max

Differential Revision: https://dagster.phacility.com/D1412