Added weather pipeline and credentials vault resource
Summary:
**Overview**

We needed to grab weather data for a set of input dates from the dark sky API. This revision creates a pipeline which performs the right api calls and uploads the dataset to gcloud storage! We also include a CredentialsVault resource to do environment variable configuration.

**Proof This works**

Run the production preset and you should see a `weather.csv` file sitting in the `dagster-scratch` bucket in gcs!

Test Plan: unit

Reviewers: #ft, max

Reviewed By: #ft, max

Subscribers: max

Differential Revision: https://dagster.phacility.com/D1452