Fix typeahead suggest token values, autoselect first completion
Summary:
This diff addresses the two issues that were making the typeahead search annoying. Rather than typing "pipeline:" and then seeing the available options, you can now type "airline_" and see "pipeline:airline_demo" right off the bat. When no chars are entered you see the available search space / keywords, but as soon as you start typing it suggests full autocompletions.

The second fix is that the typeahead autoselects the first autocompletion so you can press "Enter" to confirm your search without down-arrow first. To enter text that is NOT an autocompletion, you press esc to close the options and then press enter, but it's unclear if you'd really do that.

Test Plan: I will write tests for this component tomorrow. It's complicated enough it might be worthwhile.

Reviewers: #ft

Differential Revision: https://dagster.phacility.com/D1456