(incremental-meta-3) Make dauphin config type layer dumb objects
Summary:
This is a step in making it possible to flow through
the new meta objects into the graphql layer. In order to do this
incrementally, sometimes we will have to construct the dauphin
objects from the old config type objects, and sometimes we
will construct them from the meta objects. The meta objects
required threading around a schmea object as well to do
type lookups by key so it will be bit involved.

The config types hang out the runtime type system and
the solids so this codepath will hang around until the system
is 100% converted to using the serializable meta layer

Test Plan: BK and load dagit

Reviewers: max, alangenfeld

Reviewed By: max, alangenfeld

Differential Revision: https://dagster.phacility.com/D1475