Considerably reduce volume of type data fetched by ConfigEditor
Summary: This diff removes dead weight and duplicative data from the GraphQL query that powers autocompletion in the config editor, at the cost of slightly more explicit passing of a "list of all the inner types" to the ConfigTypeSchema component.

- We were expanding+fetching fully populated `innerTypes` of every type
- We were fetching two layers of innerTypes / ofTypes for ListConfigType

This reduces the number of lines in the GraphQL response for "Sleepy Pipeline" from 7623 to 1520, and that pipeline has very little config. Did not try with the huge dag.

Test Plan: Run Dagit tests

Reviewers: @alangenfeld, @schrockn

Subscribers: