generate empty stats object
Summary: I've gotten my instance in a state (usually in development) where a run exists in the run storage, but there are no events in the event_log.  This prevents the entire runs page to not render.

Test Plan: rendered runs page w/ borked run, saw it with NOT_STARTED status

Reviewers: alangenfeld, max

Reviewed By: max

Differential Revision: https://dagster.phacility.com/D1567