Weight DAG edges by solid name similarity, fix wrapping in sidebar
Summary:
This diff adds some basic edge ranking based on the number of underscore-separated components that appear in the connected solid names.

I also updated the sidebar to wrap the pipeline, solid and IO names because things were basically useless. Thanks to "zero width space" characters, we can wrap on the underscores in the solid names which keeps things easy to read.

Test Plan: Updated snapshots

Reviewers: schrockn, alangenfeld

Differential Revision: https://dagster.phacility.com/D1549