bypass remote launcher check
Summary:
We currently cannot launch dagit with a run launcher set because dagit is not up to pass
the sanity check at the time that the run launcher is instantiated.  Skipping this initialization
check allows us to use the same config for both dagit and the local backfill CLI, which will target
the local instance of dagit to fulfill the backfill execution.

Test Plan:
Brought up dagit with settings:
```
run_launcher:
  module: dagster_graphql.launcher
  class: RemoteDagitRunLauncher
  config:
    address: http://localhost:3333
    skip_initialization_check: True
```

Reviewers: alangenfeld, sashank, schrockn

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D1688