(python-config-type-instance-5) Moving Dict over
Summary:
Moving another one. This is the most used by far.

Note that strange bug in inference. Our multi-layer api makes
it so there are a few symbols named with high-value names like
"Dict" and it is easy to screw them up. Cost of doing things like this

Depends on D1583

Test Plan: BK

Reviewers: max, alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D1584