(python-config-type-instance-8) Type resolution in the config schema module does not accept config type classes
Summary:
This is an odd function that does a subset of the config type resolving that only the config schema
(the config schema project from the runtime types specifically). With the changes so far we can lock
it down to not accept the config classes, so we do that here.

Depends on D1586

Test Plan: BK

Reviewers: max, alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D1587