(python-config-type-instance-18) Eliminate notion of "Regular" Config Type in core. Only Regular Type was Any
Summary:
Any was the only type where this was used, so this is more
accurate. The notion of a "regular" type in the graphql layer means "Any
or Scalar" which is kind of "Scalar-ish". However not going to change
the verbiage in that layer.

Depends on D1596

Test Plan: BK. Load dagit.

Reviewers: max, alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D1597