Substantially reduce fetching required to get the entire schema
Summary:
This changes the large environment schema fetch when
the execution console is loaded to not required the querying of
innerTypes. This reduces payload for large schemas dramatically.

Also added a configTypeKey field to Field to avoid one layer
of objects in js and python layer as well as the Apollo-injected
__typename field.

Depends on D1557
Depends on D1599

Resolves https://github.com/dagster-io/dagster/issues/1959

Test Plan:
BK. Load dagit. Use config editor on large pipeline.

See schema load in around 5 secs and editor load and parse 15 seconds
later

Reviewers: alangenfeld, bengotow

Reviewed By: bengotow

Differential Revision: https://dagster.phacility.com/D1600