Remove execution_params from ScheduleDefinition
Summary:
`execution_params` really didn't make sense on `ScheduleDefinition`, so taking it out. Helps clean up code in the schedule mutation and schedule definition resolver too.

Now that we have functions that run to resolve the environment dict and tags, any error in either function would cause the scheduler page to error.

With this diff, we resolve the pipeline name and mode directly off the schedule definition (instead of getting them off execution params on the client side) and don't fail loading the page on any user errors in the schedule definition.

Test Plan: unit

Reviewers: prha, alangenfeld

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D1617