(more-config-work-6) Killing config_field
Summary:
This kills config_field across all of our decorators and
definitions in favor of a single config field that knows how to
coerce types and dictionaries into Field as well as Field itself.

Note that this has exposed a problem with API as when you pass
a type (e.g. Int) to decorator it is treated as callable
which we interpret as a decorator called without parens.

Resolves https://github.com/dagster-io/dagster/issues/1974t

Test Plan: BK

Reviewers: max, nate

Reviewed By: max

Differential Revision: https://dagster.phacility.com/D1628