Add solid subset rendering via large search bar, empty state for large DAGs
Summary:
This diff adds a large bar to the bottom of the Explore UI that allows you to type one or more solid subset expressions (eg: `++solidname` or `solidname+*`) and provides autocomplete to full solid names. The viewer renders only the subset described, making it much easier to work with large DAGs. It also adds an empty state that explains how to use the filter tool and expands on the previous iteration of the filtering logic to support expansion operators that specify recursion depth (`+` vs `++` vs `+++`)

This diff also incorporates other cleanup:
- We now center the viewport on solids you select
- We now collapse the "Modes" selector in the sidebar by default for performance reasons
- You can now copy-paste solid names and the zero-width spaces we insert to optimize display are removed

Test Plan: Run tests

Reviewers: schrockn, alangenfeld, prha

Reviewed By: alangenfeld

Subscribers: prha

Differential Revision: https://dagster.phacility.com/D1552