Address issue with pyspark context (#1980)
Summary:
We were previously using the wrong context in `SolidExecutionResult.output_value` and `SolidExecutionResult.output_values` (the `SystemPipelineExecutionContext` instead of the `SystemStepExecutionContext`. Since resources are scoped to execution steps, this meant they were not available when used to retrieve a solid result.

This updates both sites to use the correct context.

Also, we by default call `.stop()` on the Spark session when pipeline execution terminates. I've added a flag to the pyspark resource to make this configurable; if the user intends to invoke pyspark APIs outside of Dagster execution they'll need to set this to false.

The test provided gives coverage on both the correct resource / context behavior and on ensuring Spark is available outside of pipeline execution when the new config option is set.

Relatedly, this uncovered: https://github.com/dagster-io/dagster/issues/1983

Test Plan: unit

Reviewers: alangenfeld, max, schrockn

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D1643