Slightly increase solid size, allow names to wrap to two lines in DAG
Summary: This diff adds an optional `allowTwoLines` prop to our custom SVG text component that breaks on `_` and wraps to two lines.

This is not a full implementation of multiline text rendering and doesn't currently impact the intrinsic sizing rectangle used for layout. I don't think we really need one and it might be slower.

This also fixes a kind-of-major issue I discovered in which we were laying out children on SVGFlowLayoutRect and rendering them but ALSO rendering them within the parent <rect> in their pre-layout forms. This had no effect because SVG rect nodes can't actually have children, but we might get a React-side speed bump out of this fix.

Test Plan: Updated SVG snapshots

Reviewers: alangenfeld, prha, schrockn

Subscribers: