Add missing where clause to event storage
Summary:
When using `PostgresEventLogStorage`, we would see incorrect stats for runs. This diff adds a simple `where` clause to the query in `SqlEventLogStorage.get_stats_for_run`.

We didn't see this issue with `SqliteEventLogStorage` because we shard the sqlite files by `run_id`, so this where clause is not necessary in the sqlite case.

https://github.com/dagster-io/dagster/issues/2009

Test Plan: unit

Reviewers: max, schrockn

Reviewed By: max

Differential Revision: https://dagster.phacility.com/D1712