Make RunMetadataProvider less dependent on log ordering, fix dagit crash
Summary:
When many (1000s) of steps are executing in quick succession, events appear to occasionally arrive out of order, which caused the RunMetadataProvider to get into an unexpected state where `step.expectationResults` was undefined.

This diff re-writes the RunMetadataProvider to make it more straightforward and less dependent on the order that events arrive.

Test Plan: Manual

Reviewers: schrockn, alangenfeld, max

Reviewed By: schrockn, max

Differential Revision: https://dagster.phacility.com/D1708