Convert custom pandas df types to vanilla instances of RuntimeType
Summary:
Depends on D1685

Converting the generated custom data frame type to just be an instance of
RuntimeType, rather than a class mapped with a @dagster_type annotation.

I think this is more clear and it doesn't purport to be mypy compliant, which
was true before but easy to misinterpret because it was a python type.

Test Plan: BK

Reviewers: themissinghlink, alangenfeld

Reviewed By: themissinghlink

Differential Revision: https://dagster.phacility.com/D1715