Eliminate overridable type_check in custom data frame
Summary:
Having a user being able to totally override the core
type_check function in the custom df frame added a lot of complexity
and weird states the type check function can be in. I noticed this
as I was looking at this code because this type check didn't really
abide by the contract between the typecheck function and the runtime. It
threw a bunch of different exception types with low quality error
messages.

A lot of this is because the interface between user-defined typecheck
functions and our runtime is pretty poorly defined right now, and
results in quite opaque errors, rather than issues in the df type
checker itself.

Tracking that issue here: https://github.com/dagster-io/dagster/issues/2018

In this case the default behavior was not to not check the instance of
the core value at allowed. There were a number of ways that the function
could get into a state where it was throwing a CheckError or similar,
which indicates a coding error, rather than a deliberate "type check
failed successfully" condition.

Adding alex and max here because there are implications in the overall
type system, and I believe max had to add check_dagster_type to the
codebase because of some of the awkardness around all of this.

Depends on D1715

Test Plan: bk

Reviewers: themissinghlink, alangenfeld, max

Reviewed By: themissinghlink

Differential Revision: https://dagster.phacility.com/D1718