RFC: Update TypeStoragePluginRegistry to take list of tuples instead of dict
Summary:
Since we plan to make `RuntimeType`s not hashable in D1829, we can't use them as dictionary keys. This breaks multiple tests.

Currently in the codebase, we only use `RuntimeType`s as keys when constructing the dict argument to `TypeStoragePluginRegistry`. The proposed simple fix is to just turn the dict argument to a list of tuples. Alternatively, we can figure out a way to make `RuntimeType` hashable.

Test Plan: unit

Reviewers: schrockn, alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D1832