Basic composite expansion in the Dagit explore tab
Summary:
This diff adds basic expansion of composite solids to the Explore tab, addressing https://github.com/dagster-io/dagster/issues/1979. This is achieved by transforming the handles we receive from the GraphQL response and essentially "cutting" composites away and replacing them with their interior solids.

This diff adds a checkbox to enable this feature in the bottom left. Toggling it forces a re-query because we need to load solids recursively and also resets the solid query string because it's very likely your previous query is now invalid.

Test Plan: Run tests

Reviewers: prha, schrockn, alangenfeld

Differential Revision: https://dagster.phacility.com/D1860