Bay Bikes Weather Pipeline refactor
Summary:
**Summary**

This revision is an incremental refactor to getting the bay bikes pipeline to be more real world and work with our partitions API's and schedules. Before the pipeline was downloading a csv file and inserting a row (just because it was expedient), but the partition backfill API necessitated a pipeline that was transactionally consistent, so we ended up moving to a database. As I was developing, I realized that I too needed the db related resources in the airlines example so I moved those resources out to a `dagster_components` directory (heavilly inspired by *cough* react *cough*) (*cough* elementl data design system *cough*).

**changelog**

- Radically changes the `extract_daily_weather_data_pipeline`
- Change `download_weather_report_from_weather_api` solid to just return a dataframe for one day.
- Create `insert_row_into_table` solid
- Moved db resources out to a `dagster_components` directory to encourage code reuse.
- Add solid tests for `download_weather_report_from_weather_api`
- Added schedules for pipeline and checked backfill api worked!

Test Plan: unit

Reviewers: prha, sashank, alangenfeld

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D1874