(4/n scheduler-refactor) Move ScheduleStorage to DagsterInstance
Summary:
Previously, the `ScheduleStorage` for a `Scheduler` was hardcoded to be `FilesystemScheduleStorage` inside `SchedulerHandle` (line 72 of scheduler.py). `SchedulerHandle` is initialized when we load Dagit.

This diff moves `ScheduleStorage` to `DagsterInstance`, and removes the `_schedule_storage` private variable on `SchedulerHandle`. Instead, the schedule storage must be passed directly to the methods of `SchedulerHandle`.

Test Plan: unit

Reviewers: alangenfeld, prha

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D1911