(type-system-changes-9) PythonObjectType --> PythonObjectDagsterType. Eliminate define_python_dagster_type
Summary:
So this is RFC-ish. I've been struggling naming the idea of
creating a dagster type where all you want to do is an instance of
check. Here I'm proposing that we just use the class
PythonObjectDagsterType and eliminate the define_python_object_type.

This will make it clear that it much more similar to the DagsterType
and doesn't involve any sort of global registry

Depends on D1906

Test Plan: BK

Reviewers: max, alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D1930