(type-system-changes-10) dagster_type --> usable_as_dagster_type; introduce make_python_type_usable_as_dagster_type
Summary:
Our long national nightmare of dagster type naming is about to
come to end -- or at least a temporary cessation. There will, at last,
be peace in our time.

The goal of this rename is to "downgrade" the @dagster_type decorator
from our core workhouse type creation function to a more explicit, less
foregrounded equivalent, @usable_as_dagster_type.

This name indicates that this python class can be passed to
anywhere that would expect a dagster type, such as an argument
to an InputDefinition or an OutputDefinition

Additionally we expose make_python_type_usable_as_dagster_type to
allow the user to make *existing* classes usable and correspond
to a dagster type. This would be used, as an example, to import
a class from a library you do not control, and make it usable
in dagster type contexts.

Test Plan: BK

Reviewers: alangenfeld, max

Reviewed By: max

Differential Revision: https://dagster.phacility.com/D1933