Don't show schedule attempt error when using the run launcher
Summary:
We give `ScheduleAttempt` a status based on the typename of the graphql response from `startScheduledExecution`.

Previously, we would assign a status of `SUCCESS` if `typename == 'StartPipelineExecutionSuccess'`. Since we changed the behavior to use the run launcher for a scheduled run when it is defined, we need to add `'LaunchPipelineExecutionSuccess'` as a success state as well.

Test Plan: Use run launcher + scheduler, verify error isn't shown.

Reviewers: prha, themissinghlink

Reviewed By: themissinghlink

Differential Revision: https://dagster.phacility.com/D1936