refactor PandasColumn API
Summary: This revision is a follow up to https://dagster.phacility.com/D1971. It refactors the `PandasColumn` to use `staticmethods` instead of `classmethods` for compositions to reduce the surface area of the API. It also gets rid of `add_configurable_constraints` in favor of a `_construct_keyword_constraints` helper and gets rid of the pseudo registries to be more explicit about constraint construction. It also changes the exists, kwarg to non_nullable to be more explicit about what constraint is actually being applied (and it doesn't conflict with the exists composition).

Test Plan: unit

Reviewers: schrockn

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D1975