Move playground back into pipeline scope, clean up route props
Summary: This is a small experiment that moves the playground tabs back within a pipeline selection. I think that this feels a bit more natural, and it also eliminates the need to "clean" the session as pipelines appear and disappear in the repository.

I also took the opportunity to clean up a few of our route components which weren't using RouteComponentProps and were re-defining built in props like `history` and `match`.

Test Plan: Run snapshot tests

Reviewers: prha, schrockn, alangenfeld

Subscribers: