Generate screenshots for intro tutorial
Summary:
Run `yarn screenshots` from the `/docs` folder.

The script is very verbose and could perhaps be refactored, but there are a lot of touchy scope issues at the intersection of async functions and jest specs, so rather than fiddle with that for too long I figure this is a reasonable starting place. There is some brittleness from the use of XPath selectors -- we should also eventually get to a place where there are semantically meaningful, consistent ids on our UI elements -- but I don't think any of that should block the immediate usefulness of this, which brings screenshot generation down from a ~1 day, error-prone, manual process to ~6 minutes.

Test Plan: Manual

Reviewers: sashank, prha, nate

Reviewed By: sashank

Subscribers: bengotow, catherinewu

Differential Revision: https://dagster.phacility.com/D2143