Fix issue with schedules in test image
Summary:
Dagit was crash looping in our Helm tests. This diff fixes that issue.

We previously would've caught this in tests, but we no longer wait for Dagit to become available, as it would require first rebuilding Dagit before deploying. We removed that step because it adds substantial time (2-3+ mins) to the critical path of the build

Test Plan: buildkite

Reviewers: alangenfeld, max

Reviewed By: max

Differential Revision: https://dagster.phacility.com/D2167