add execute_pipeline_with_mode and tests
Summary: I had a situation where I was testing and using an environment dict that I was transforming a bunch but still needed to use modes to utilize testing resources. I realized that there is no way to pass the mode in without creating your own RunConfig which seems a bit messy. This revision just adds some sugar to make this possible.

Test Plan: unit

Reviewers: max, alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D2196