Support for markers in the Gaant timeline visualization
Summary:
This diff adds Gaant visualization support for markers defined in EngineEvents. Right now the visualization is fairly minimal. Markers with no associated step are shown on the first row of the visualization. Markers with an associated step are shown on the same row as the step. The markers are semitransparent purple with a bit of a "box-and-whiskers" appearance. They currently are not selectable, and selecting a step highlights it and it's execution time non-inclusive of the step-associated markers. I'm not sure if this is the ideal behavior but I think it'll get us going!

You can hover over the markers to see the strings that are sent down in the EngineEvent markerStart and markerEnd fields. Right now for step markers, these strings are prefixed with the step name to make them all unique.

Test Plan: Updated snapshots

Reviewers: prha, alangenfeld, schrockn

Differential Revision: https://dagster.phacility.com/D2210