(snap-backed-graphql-7) Back all dependency-related things via snap
Summary:
This uses the pipeline and dependency indexes to back all
dependency-related graphql types.

This is the most thorny part of the graphql schema and I believe it
should be relatively straightforward from here to convert the rest
of the schema to be snapshot backed.

Depends on D2310

finallydone

Test Plan: BK

Reviewers: alangenfeld, sashank, themissinghlink

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D2311