Create prod mode and preset for airline demo
Summary:
This more or less works with live Redshift/EMR, will likely have a few more fixes but this is getting to be a big diff. That one `# FIXME` to make this work on EMR ended up being a bunch of things...

Shouldn't be anything complicated/controversial here, but lmk and I can break this up for smaller reviews if necessary.

Test Plan: unit and manual with EMR/Redshift

Reviewers: max, schrockn

Reviewed By: max

Differential Revision: https://dagster.phacility.com/D2318