(stable-pipeline-snapshot-1) Use json rep default value in config type key
Summary:
This does a couple things:

  1) changes the config system to have a codepath that expands defaults
without doing a post-process step. This is necessary to allow default
values that have a nested enum to be serializable

  2) Incorporate the json serialized default value into the config key.

This changes the config type key to incorporate a default
value hash instead of a string representation of the default value.

In order to do this has to change the default expansion to not to
post_processing.

Test Plan: BK

Reviewers: alangenfeld, themissinghlink

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D2393