(container-supported-dagit-1) - Implement DagsterContainerGraphQLContext and hook into pipelines_or_error
Summary:
This revision introduces a DagsterContainerGraphQLContext and renames the `DagsterGraphQLContext` to `DagsterHandleGraphQLContext` to be more explicit. Then to show how it ought to be used, it hooks it into the `pipelines_or_error` resolver.

I didn't add exception handling to presets yet, because this is only triggered by pipelineOrError resolvers and that would have blown up the revision (refactoring `get_pipeline_def_from_selector` is gonna be its own revision coming up next).

Test Plan: unit

Reviewers: schrockn, alangenfeld

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D2461