(container-supported-dagit)-support-used-solids
Summary: This revision introduces the RepositoryIndex which indexes and holds all of the pipeline indices on context instantiation. I went the pre-computation route refactored all downstream code so that instead of constantly constructing PipelineIndices, they just grab the pipeline index from the repository index held by the context. Then I used it to refactor usedSolid and usedSolids which was the major motivator here because that codepath has had some perf issues in the past. This cleans things up pretty majorly which is awesome and limits the instance forking to solely the fetch pipeline resolvers.

Test Plan: unit

Reviewers: schrockn, alangenfeld

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D2512