display failure metadata in expanded error modal
Summary:
We were not adding metadata to the PythonErrorInfo view.

Long-term, we need to not duplicate the error rendering logic in two places, but that
might be a more invasive process.  Taking the easy wins now, and tracking the larger
effort in https://github.com/dagster-io/dagster/issues/2371.

Test Plan: Created `Failure` with metadata, saw metadata rendered in the expanded view

Reviewers: sashank, schrockn, alangenfeld

Reviewed By: alangenfeld

Subscribers: alangenfeld

Differential Revision: https://dagster.phacility.com/D2543