(container-execution-preparation) - Add config type deserialization
Summary: We have a few resolvers which will be constructing environment_config's for use in playground validation. Since we want to back ernvironment config schema generation with the pipeline index, we are going to need the ability to deserialize the pipeline index ConfigTypeSnaps to actual config types.

Test Plan: unit

Reviewers: schrockn, alangenfeld

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D2540