fix config type snapshot tests
Summary: It looks like pytest snapshot just serializes python objects to GenricRepr strings which generates a plaintext with a hex that differs from OS to OS. This isn't testing anything. As a result, we are now taking the config type objects, serializing them to namedtuples and snapshotting that output. This should fix the test and actually produce a reliable test.

Test Plan: Unit

Reviewers: max, nate

Reviewed By: max

Differential Revision: https://dagster.phacility.com/D2560