Add a fourth state to the Run Preview for nodes that are required but missing
Summary:
This is a small diff that fetches a bit of additional data from validation errors to create a fourth state in the run preview showing solid blocks that need to be defined but are currently missing. Previously these were shown in light blue, so they looked fine / ready when they still needed to be defined.

{F117504}

Test Plan: Run tests

Reviewers: prha, alangenfeld, schrockn

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D2537