Allow reimports of dagstermill and multiple calls to get_context at runtime
Summary: Fixes an issue Milos Tomic encountered when reimporting dagstermill and getting an interactive context (`dagstermill.get_context()`) in a cell after the injected parameters cell of a dagstermill notebook. This was overwriting the instance of the noninteractive context on the manager. Now we cache the noninteractive context for further calls to `get_context`.

Test Plan: Unit

Reviewers: prha, alangenfeld, nate

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D2569