(new-dagster-type_api-1) New GraphQL API for fetch dagster types
Summary:
Think of this more as a proposal. I'm not a
huge fan of "double-layering" the "OrError" types here.

But this ends up scoping the runtimeTypeOrError field
to any pipeline snapshot rather than having it at global
scope.

An alternative would be to change the root call to take
a snapshotID and force the client to manage snapshot ids
even in cases where you can querying by pipeline name

Test Plan: BK

Reviewers: alangenfeld, bengotow

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D2595