[WIP] Initial support for clicking through to pipeline snapshots from Runs list
Summary:
{F119596}

A few thoughts about this:

It's a bit annoying to need to switch the root query in GraphQL to retrieve data from two different APIs. I abstracted the difference in queries away in the `PipelineExplorerRoot.tsx` file using two small helper components and I'm pretty pleased with that solution (the entire explorer just deals with an IPipelineSnapshot shape now), but we need to do it in a few more places and my current solution is fairly code-heavy.

The problem is that we need to swap out the top line of 3-4 queries like this (`pipeline(params: { name: $pipeline })`), but everything inside stays the same. If we'd written our own GraphQL layer I'd be tempted to just string-substitute the right query into the top line of the string and fire it off, but I don't think Apollo will let us do that. We can't IF / ELSE the query within a React component because the `useQuery` hooks aren't designed for the query to change, only the variables and hooks cannot be conditionally defined in components.

```

export const SIDEBAR_TABBED_CONTAINER_SOLID_QUERY = gql`
  query SidebarTabbedContainerSolidQuery(
    $pipeline: String!
    $handleID: String!
  ) {
    pipeline(params: { name: $pipeline }) {   << need to change only this line in several more queries for sidebar
      name
      solidHandle(handleID: $handleID) {
        solid {
          ...SidebarSolidInvocationFragment

          definition {
            __typename
            ...SidebarSolidDefinitionFragment
          }
        }
      }
    }
  }
  ${SidebarSolidInvocation.fragments.SidebarSolidInvocationFragment}
  ${SidebarSolidDefinition.fragments.SidebarSolidDefinitionFragment}
`;

```

It seems like it'd be really nice if the current state WAS a snapshot with a named alias. So I could just say `snapshotID || "current"`, and always use the same query to retrieve the data.

Alternatively something like `pipeline(params: { name: String!, snapshotID: String! | null }) { }` would be nice too.

Test Plan: Updated snapshots but probably need to add a new one

Reviewers: schrockn

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D2613