Create chart.js-based run status table view
Summary:
The home-grown table view for runs was really broken for large time ranges.
This diff re-implements this using the canvas-based chart.js.  It hacks the chart.js
datasets by using a Line graph, but not showing a line, and representing each type
of run status as a different dataset, while customizing the tooltip, point representation,
and onClick behavior

{F123650}

Test Plan: bk

Reviewers: sashank

Reviewed By: sashank

Differential Revision: https://dagster.phacility.com/D2654