fix out-of-order issues in multiproc test
Summary:
test_separate_sub_dags was failing due to out-of-order failures and now is failing because
of intermittent file permission errors.  Suspect it's because D2671 switched it to open the file
to check for step completion instead of just checking file existence / file size.

To deal with this we change how the tests works to keep the spirit of the test intact but deal with this ordering problem

resolves #2404

Test Plan:
https://dev.azure.com/elementl/dagster/_build/results?buildId=7088&view=results

https://dev.azure.com/elementl/dagster/_build/results?buildId=7089&view=results

Reviewers: max, prha

Reviewed By: prha

Subscribers: yuhan

Differential Revision: https://dagster.phacility.com/D2678