(active-repository-data-6) Add ExternalPipeline. Use it to power DauphinPipeline
Summary:
This PR makes preset values available over a serializable
interface. It orders the ExternalPipeline object, which is in-memory on
the dagit side and composes a PipelineIndex and the ActivePipelineData.
It is the object that represents a pipeline that is active in another
process/environnent that dagit knows about

Test Plan: BK

Reviewers: alangenfeld, sashank

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D2690