(external-config-schema-4.5) Changing post process API contract
Summary:
This buttons up the post process API contract for config types
so that we have an expected exception type that communicates failures
which are expected to occur, to distinguish them from errors which
are internal errors.

Depends on D2751

Test Plan: BK. These errors never worked properly in dagit.

Reviewers: sashank, alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D2753