(external-config-schema-4.8) Eliminate Path from config type system
Summary:
Treat this as RFC-ish, but I think we should do it. The Path
type used to have a purpose: it no longer does not. The only behavior
difference you get if you use Path is config scaffolding is different.

I think it just adds confusion and we should just do strings instead
of Paths.

Depends on D2754

Test Plan: BK

Reviewers: sashank, alangenfeld, nate, max

Reviewed By: alangenfeld, nate

Differential Revision: https://dagster.phacility.com/D2755