Stop spurious default usage
Summary:
I've been seeing this in the codebase a lot and I think its
important we don't do this. Meaning defaults where we don't need them.
With long parameter lists it is *incredibly* easy to miss one. In this
case this is a private helper with only two callsites, and all params
from the callers in those need to be passed through.

Our "default" position should be to *not* have defaults for non-public
APIs. If we feel like we need defaults for testing convience, we should
have explicit test functions whose sole purpose is default application,
so that we ensure prod codepaths are correct

Test Plan: BK

Reviewers: max, alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D2851