Kill dagster.Path
Summary:
We killed Path in the config system but left it in the dagster
type system, which was a mistake. This just moves forward with totally
eliminating Path.

Alternatives would be including a tombstone object and then erroring on
it with a specific error message, because right now the error message
can be a bit rough: https://dagster.slack.com/archives/CCCR6P2UR/p1589121638025100

Test Plan: BK

Reviewers: max, alangenfeld, prha

Reviewed By: max

Differential Revision: https://dagster.phacility.com/D2854