Fix external pipeline subselection bug
Summary:
Previously, when we called `create_execution_plan_index` with an external pipeline that was created from a solid subset, the pipeline definition that we retrieved for passing to `create_execution_plan` was the full pipeline, not the pipeline w/ subset. This is because we didn't persist the `solid_subset` information anywhere, and just used the `pipeline_name` to retrieve the pipeline definition.

This diff stores the `solid_subset` information on `ExternalPipeline` so we can re-use it to retrieve the pipeline w/ subset

Test Plan:
1. Load pipeline with multiple solids that require config
2. Select solid subset in Dagit
3. Delete extraneous config for solids that are not in the subset
4. Verify that there is no error

Reviewers: schrockn, alangenfeld, max

Reviewed By: schrockn, alangenfeld

Differential Revision: https://dagster.phacility.com/D2870