(host-representation-1) Move ExternalPipeline into dagster.core.host_representation
Summary:
This is the beginning of an extended series of diffs to
eliminate the Index classes that wrap snapshots.

In its stead, we will introduce a new subpackage called
host_representation, whose purpose is to define classes that
1) host processes such as dagit can use to represent definitions
that are loadable into memory and 2) historical artifacts
that we have persisted in the instances.

Test Plan: BK

Reviewers: sashank, alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D2898