Initial version of wider, pipeline-centric left sidebar
Summary:
This diff replaces the Dagit sidebar with one that clearly separates "instance"-scoped concepts (assets, schedules) from "environment" scoped pipelines and solids. The new design is also more pipeline-centric and shows the environments pipelines directly in the sidebar. We may end up collapsing this back into a "narrow" bar or having it slide out on hover, etc., but we wanted to establish the new design first and then revisit whether it'd be possible to shrink.

This is the first of a couple major changes we want to make to the Dagit UI to reflect the new pipeline focus - the plan is to swap out the "Definition / Playground" options in the sidebar (sort of a sub-selection right now) for a set of top tabs, but that change involves updating a lot of routes / URLs and I think it'll be easier to review separately.

Todo Soon:
- Where should environment name come from? Need a small GraphQL API for this until we have multi-env support? (OR: is there an agreeable default we can use for now?)

Lower priority improvements up for grabs:
- Implement keyboard up/down + enter in the inline search box
- Make the left nav column resizable by dragging the right edge
- Unify the empty states of the Runs / Schedules / Assets pages
    + Runs has header, schedules doesn't
    + Assets background is a different color, centered content is further down

{F140554}

Test Plan: Update snapshot tests

Reviewers: alangenfeld, schrockn, prha

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D2967