Remove repository definition from ScheduleExecutionContext
Summary:
This diff removes the repository definition from the schedule execution context. This was initially added as a hack for the backfilling scheduler, but is unecessary because schedule definitions and repository definitions are co-located now.

This lets use remove a few more calls to `get_repository_definition()` in `dagster-graphql`

Test Plan: unit

Reviewers: schrockn, alangenfeld

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D2972