Formalize dagit "pipeline scope" with "Overview", "Definition", "Playground" tabs
Summary:
This diff adds a top nav to the playground and explore tabs and creates a placeholder for the Overview tab @prha is working on. I reworked this to avoid changing the URLs significantly, and was able to use a few React router features to pull in the context we need to maintain tab state across pipeline selection changes.

The pipeline subset is still managed in the same way via a special `pipeline:subset` selector syntax that is used in the URL:
eg: http://localhost:3000/pipeline/airline_demo_ingest_pipeline:join_q2_data++/

I think that we can swap this out for a query string approach, but we have a set of helpers in place that make this work nicely and I think it'd make sense to pull them out in a separate diff.

I also addressed @prha's comments from the previous diff (moving things to `nav` folder, which definitely makes sense now that there is also PipelineNav). The sub-selection in the sidebar is gone now that we have the top tabs.

{F141474}
{F141473}

Test Plan: Updated snapshots

Reviewers: alangenfeld, schrockn, prha

Reviewed By: schrockn, prha

Differential Revision: https://dagster.phacility.com/D2994