"Stream" events from S3 to plan process
Summary:
This is the "read" side of #2443.  A background thread in the plan process periodically reads an object in S3, checks whether it contains any events that the plan process hasn't received so far, and, if it does, yields them.

This is a little tricker than the write side, because we have two interleaved pollings happening: polling EMR to determine whether the step is done and polling S3 to get events.

Test Plan: added tests, ran test_do_it_live_emr and observed events streaming in.

Reviewers: nate, schrockn

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D2990