Add CliApiRunLauncher [redux]
Summary:
Putting this up again because I want this to go through
proper dev and qa post rebase.

Adds the CliApiRunLauncher. This is largely insprired
(copied) from the PipelineExecutionManager, which will
eventually be deleted. This does the same monitoring and
running of pipelines as that, but uses python.threading
instead of gevent, subprocess.pOpen instead of multiprocessing, etc.

Test Plan: BK

Reviewers: alangenfeld, max, sashank

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D3034