"Stream" events from EMR to S3
Summary:
This is half of #2443 - a background thread in the EMR process periodically rewrites an events file on S3, containing all events that have been yielded by the step so far.

The last time I wrote multithreaded code was in Java, so possible I'm not following the right conventions.  I avoided using asyncio because there doesn't seem to be a good compatibility story with python 2.

Even though this only changes remote side, it should still be compatible with what the plan process side expects - in addition to all the events being written at the end of the step, subsets of them are written during the step's execution.

Test Plan: added tests, run test_pyspark::do_it_live_emr

Reviewers: nate, schrockn, alangenfeld

Reviewed By: nate

Differential Revision: https://dagster.phacility.com/D2968