(run-hijacking-9) Add tests for the test suite itself
Summary:
As this is getting more complex, it's going to become
more easy to screw up. So these are a bunch of tests and new conventions
for the new test suite itself.

It puts some guard rails in place to ensure that marks are setup correctly.

Top level static methods on GraphQLTestInstances, GraphQLTestEnvironments,
and GraphQLTestExecutionManagers must be in the MARK_MAP. In turn
these marks are used to filter the context variants that contain those things
to build convienent sets of context variants to use in different test suites.
This should encourage us to keep these up-to-date and accurate.

We then test those prefab sets of variants to ensure that they have the properties
that we expected. e.g. all variants in
GraphQLContextVariant.all_variants_with_legacy_execution_manager should in fact
have execution_manager set.

I also re-recorded all snapshots, hence the thrash there.

Depends on D3099

Test Plan: BK

Reviewers: alangenfeld, max

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D3100