(hijack-cli-api-all-things-8) Make start_pipeline_execution_for_created_run execute directly in process
Summary:
This is quite a convuluted codepath which should be getting
better soon. We were using the graphql API to implement the bodies of
run launchers using this API. So it should be synchronously
executing always. But this was dispatching through the execution
manager, which actually doesn't make sense. This diff changes
this code path just to call execute_run directly.

In the hopefully-not-too-long term, this entire functionality
should be removed from dagster-graphql and instead be invoked
using the dagster cli api (or its successor).

In the short term we need this in order to remove execution manager.

Depends on D3147

Test Plan: BK

Reviewers: alangenfeld, max, sashank

Reviewed By: alangenfeld

Subscribers: alangenfeld

Differential Revision: https://dagster.phacility.com/D3150