(dagit-out-of-process-3) Add validate external pipeline command
Summary:
This diff adds a command that lets us verify using the CLI API whether a subset for a give pipeline produces an invalid subset or throws an `DagsterInvalidSubetError`exception.

We also re-wrap a `DagsterInvalidDefinitionError` thrown while creating a pipeline subset as a `DagsterInvalidSubsetError`. We were previously doing this in `dagster_graphql` but are now pushing it to a lower level.

Test Plan: unit

Reviewers: alangenfeld, schrockn

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D3222