Rename "definition" to "attribute" in workspace loading pathways
Summary:
Renames "definition" to "attribute" in workspace.yaml and
its command line equivalents. @alangenfeld gave the feedback that
definition was confusing because we use the word definition in other
ways in the system, namely the user-defined in-memory objects like
PipeilneDefinition and friends.

"Attribute" is good here because it means that we are looking for a
given attribute within a python module we are looking at.

Test Plan: BK

Reviewers: alangenfeld, max

Reviewed By: alangenfeld

Subscribers: alangenfeld

Differential Revision: https://dagster.phacility.com/D3341