Add repo switcher to the sidebar that reads/writes DagsterRepositoryContext
Summary:
I also modified the repo context a bit so that it is never `undefined` by manually passing the current selection and available options into the sidebar (which I think should be the only thing rendered when the selected repo is invalid or the repos are still loading.)

After playing with it a bit more, I am finding the repo + repo location pair to be a bit awkward because the pair define a repo so it's easy to up with `repo.repository.name`. I changed the TS interface to be called `DagsterRepositoryOption` (repo + repo location) but I'm not totally satisfied with that naming either. I think that from a front-end perspective it may be nicer to have an array of repositories that each have a nested `repositoryLocation`. Even if the `repositoryLocation` on several of the options is the same and we group them client-side, it means that you have a single object to pass around?

I could also see exposing an opaque `repository.id` which quietly combines `repo name + repo location name`, but I think for people using the CLI / graphql interface manually, having to provide the tuple does reinforce the correct mental model of the system :-/

{F154907}

Test Plan: Run tests

Reviewers: schrockn, alangenfeld, bengotow

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D3292