Only pass pipeline run id to execute run cli api
Summary:
When dealing with large config, we run into a "OSError: [Errno 7] Argument list too long" error since we are tyring to serialize and pass the full pipeline run to the cli api.

Instead, we just pass the run id and retrieve the pipeline run in the user process, just like we did with the k8s run launcher.

Test Plan: unit

Reviewers: catherinewu, alangenfeld, schrockn

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D3350