(snapshot-perf-2) Precompute snapshot ID cache
Summary:
We spend a lot of time recomputing snapshot ids over and over again as we recompute it when
ever we get the full external pipeline out of the context. This caches that value to that
we don't recompute that many times on pages like the overview page.

Accomplished this by adding a class PipelineSnapshotWithID and forcing it to be passed into PipelineIndex instead of re-doing the computation everytime it is constructed.

Depends on D3504

Test Plan: BK. Click dagit. See profiler images

Reviewers: prha

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D3507