scheduler K8sRunLauncher workaround
Summary: Described in comment in code, this is an issue faced when `launch_scheduled_execution` fires a `K8sRunLauncher` which currently goes through the GraphQL call `executeRunInProcess` to perform the execution. Since this is going from a "user process" to a "host process" - we are lacking information about the host structure, in particular the `repository_location_name` that the repository we care about is hosted under.

Test Plan: added test case

Reviewers: schrockn, prha, alangenfeld

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D3517