fix the check to automatically stop the example backfill schedules
Summary:
was getting schedule attempt errors because the instance api for
`stop_schedule_and_update_storage_state` changed to take origin id.

Could not resolve origin id because we are evaluating outside of the named repo
context and all we have is the instance. Instead, we continually skip until we
find a run to execute.

Also slightly tweaked the backfill logic, so that it will try to kick off runs
for partitions that have no successful runs (as opposed to no runs at all).

Test Plan: check that schedules ran

Reviewers: sashank, nate

Reviewed By: sashank

Differential Revision: https://dagster.phacility.com/D3557