update schedule page to show only last tick
Summary:
Right now, the tick stats are misleading because we forced users to clear schedule storage
when they upgraded to 0.8.0.  Now the tick counts are way off from the runs associated with the
schedule.

In reality, we only really care about the last attempt.  So this diff removes the tick table,
coalesces a lot of the query complexity, and fixes some other schedule bugs.

{F165463}

Test Plan: Viewed schedules page, schedule page

Reviewers: sashank, nate

Reviewed By: sashank

Differential Revision: https://dagster.phacility.com/D3546