Add location reload button when supported, keyboard nav in sidebar
Summary:
This re-enables the reload button and wires it up to the new mutation for reloading a repository location.

I also attached the Option-P keyboard shortcut to the search bar and made it so the up/down arrow keys allow you to move through the list of pipelines below. This should allow full keyboard navigation of the left sidebar.

Test Plan: Run tests

Reviewers: schrockn, bengotow

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D3569