Fix injection when there is no tagged parameters cell
Summary: Previously, if there was no tagged parameters cell, we were injecting into the *second* cell of the notebook instead of the *first*. This meant that if the user did anything that required the reconstituted context in the first cell of the notebook, such as importing dagstermill and then yielding a result, it would fail. Adds a test case.

Test Plan: Unit

Reviewers: leoeer, alangenfeld, prha

Reviewed By: leoeer

Differential Revision: https://dagster.phacility.com/D3623