(celery-k8s-refactor-2) Separate global app and task instances into own file
Summary:
Celery requires that there is a module that one
points to load the "app" when invoking it from the command line.

This diff separates that out to intos our single-purpose module.
This is cleaner and that global "app" object will never
be created unless that module is explicitly included.

Test Plan: BK

Reviewers: max, alangenfeld, nate, catherinewu

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D3646