(build-image-improvement-2) Add slimmer bootstrap test builder image for faster cold builds
Summary:
Right now we use the integration image to build the test images
that later tests used. We don't need an image this heavy to do that.
This adds a new slimmer image (652MB versus 6.41GB) to do the test image
builder only.

This reduces time spent in this build step by approximately 2 minutes
when the integration image is not cached (e.g. on a cold cluster)

Depends on D3716

Test Plan:
BK

before:

{F184145}

after:

{F184147}

Reviewers: nate

Reviewed By: nate

Differential Revision: https://dagster.phacility.com/D3717