[docs] Remove landing page and /docs path
Summary:
Previously, the docs site had three top level paths: `/_modules`, `/apidocs`, and `/docs`. All of the manually written content lived under the `/docs` folder. This diff hoists all the content to the top level, and removes the `/docs` directory. Additionally, it renders the sidebar when displaying the homepage. Now, the sidebar is always visable, and there is no longer a separate landing page.

This is still a WIP, since there are several links to `/docs/...` paths within the codebase.

{F184452}

Test Plan: bk

Reviewers: sandyryza, nate, yuhan

Reviewed By: sandyryza

Differential Revision: https://dagster.phacility.com/D3725