Add missing step to pyenv setup in contributor docs
Summary:
This is something I ran into when I was getting setup this morning. The step here is needed for pyenv to work: https://github.com/pyenv/pyenv-virtualenv/#installing-with-homebrew-for-macos-users

There may be a better way in Markdown to have the text stay roughly inline with the bullet points above?

Test Plan: Build docs site, view contributes page (see attached screenshot)

Reviewers: alangenfeld, sandyryza, nate

Reviewed By: nate

Differential Revision: https://dagster.phacility.com/D3772