Coverage-specific docker image
Summary:
The coverage step is by definition on the critical path for
builds so its worth optimizing it a bit.

This docker image is 427MB versus the 6 GB integration images. In cases
where the docker image is uncached this saves about 1.5 minutes (the new
one downloads and extracts in 8-10 seconds)

This also installs git, ruby and required python modules  etc in the parent image
which shaves ~20 seconds in the cached image case..

Test Plan: BK

Reviewers: nate, max

Reviewed By: nate

Differential Revision: https://dagster.phacility.com/D3800