dont create Executor in execute_plan
Summary:
When doing sub plan execution, we don't use the executor so we shouldn't create it. The main thing this avoids is unnecessary failing of checks the executor may do despite not being relevant.

Turns the execution context creation flow in to a class setup for differing overrides in the plan and pipeline execution cases.

depends on D3806

Test Plan: added test

Reviewers: schrockn, sandyryza, prha, max, dgibson

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D3667