Pass in DagsterInstance to RunLaunchers before launch_run
Summary: Right now the only time that a RunLauncher gets access to the DagsterInstance is the first time a run is launched, making it impossible for RunLauncher to terminate a run that was previously started in a different session. This diff adds a separate method for setting the DagsterInstance that it calls immediately on startup. A subsequent diff will take advantage of this method to let you terminate a run that was previously started in a different process

Test Plan:
BK tests
Sanity check that the cli api run launcher in dagit still works
Are there any other manual test steps that reviewers would reocmmend since this touches several different run launchers?

Reviewers: alangenfeld, schrockn, max, nate

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D3817