Tutorial docs copy suggestions and bugfixes
Summary:
I did a pass through the tutorials as a person relatively new to the product, looking for any bugs or places where I got confused. These suggestions are the result of that pass.

There were a few places where the code examples were offset incorrectly, and I tweaked some wording throughout. The  wording suggestions in particular are completely optional and feel free to push back on any of them.

I'll put a couple other suggestions in a quip for things that aren't really expressible as a diff.

Test Plan: Read through each page of the tutorials again, scan each change to make sure

Reviewers: sashank, sandyryza, yuhan, nate

Reviewed By: sandyryza

Differential Revision: https://dagster.phacility.com/D3831