Don't call DagsterGraphQLContext.drain_outstanding_executions outside of tests
Summary:
I'm in the process of making it so that the run launcher stores in-progress launched runs on disk and loads them back up again when dagster starts up. This means that calling run_launcher.join is potentially problematic if other background processes call it when they weren't supposed to (whereas before they would have been harmless since any run launcher state was just stored in memory).

Right now every time we execute a query against dagster-graphql, we call drain_outstanding_executions, which calls run_launcher.join. The problem is that execute_query is also called  from places like dagster-dask workers that have nothing to do with the run launcher,

Moreover, after talking with alex it seems like this call may not be needed either in the common case where you call dagster-graphql to launch a pipeline run. In Alex's words,

"it may just be a left over artifact of how things used to work with execution manager
if you do a dagster-graphql invocation to launch a pipeline run, it should just exit when the launch is done which is once control is handed over to the executing process"

So this diff just removes the call altogether

Test Plan:
BK
Launch a pipeline run from dagit, verify that it goes through execute_query and still executes successfully
No more dagster-dask test failure on D3850 when a dask worker tries to clear out runs in the middle of the run because it executes a graphql query

Reviewers: alangenfeld, schrockn

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D3872