Make config editor help a split pane to avoid overlapping text #2681
Summary:
It seems the config editor's help panel is typically pretty large now that people are writing inline documentation for config, and it currently overlaps the text editor.

This diff switches it to a panel using our split pane component, and adds a button above it that toggles it open / closed which might be convenient for smaller screens.

I also refactored a few things I noticed were broken in the split panel component. Now that we have a sidebar, storing the panel setting in "vw" and "vh"(viewport width and viewport height) isn't quite right and means that "full width" might be 90 or 80 and not 100. Another small change also prevents the divider from being hidden / impossible to drag offscreen.

{F201887}

Test Plan: Run tests

Reviewers: alangenfeld, prha, schrockn

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D3886