Ensure the repository_code_pointer_dict is always keyed by repository name, eliminate RepositoryHandle.repository_key
Summary:
Right now RepositoryHandle has repository_name and repository_key. The former is the name of the repo, the latter is how the repo location keys it. The problem with this is that it makes it impossible to create an ExternalRepository unless you have everything you need to construct the full RepositoryLocation (an example where we do not currently have this information is the scheduler, so this is preventing us from using ExternalRepository there currently)

The only place where repository_key and repository_name are ever different is the specific case where you specify an attribute in your load_from (in this case, repository_key is the attribute, not the repository name). This diff eliminates that special case by allowing an attribute in get_loadable_targets (much like the grpc server was already doing).

This has the nice side effect of sharing more code between the grpc and non-grpc codepaths, and simplifying the workspace loading logic.

The only downside here than I can see is that we now have to go through the CLI API for that specific case (where you specify at attribute in your workspace yaml). That seems like a fair trade to me.

Depends on D4009

Test Plan: BK

Reviewers: alangenfeld, schrockn, sashank

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D4010