[celery-k8s] schedule job name only once
Summary:
This addresses a user encountered error. When 'task_acks_late: true' is set in celery config, re-deploying celery can cause the same celery task to be run more than once, which means that we will try to create more than one K8s Job with the same name. This results in a 409 Conflict error from the kubernetes API. This diff handles the error by aborting the creation of the the K8s Job if another Job with the same name exists (job names are unique in dagster). If the error is unexpected (and not the 409 ^), report an engine event and return.

Future: We could consider handling this situation by creating a job with a new / different job name if there is a use case for it.

Test Plan: manual

Reviewers: alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D4016