Add new params to get_external_schedule_execution_data to support using it in the scheduler
Summary: We need two things in the scheduler that aren't currently exposed over the CLI API or GRPC - the tags, and whether the scheduler can execute. This adds some config to the corresponding API to support that. We still want to support the existing usage from graphql that doesn't need to bother calling should_execute, so added a config param as well that lets you indicate whether you care about checking if it can execute first.

Test Plan: BK

Reviewers: alangenfeld, sashank, max

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D4029