(instance-scheduler-1) Update sidebar navigation for schedules
Summary:
In preparation for the instance level scheduler page, this diff replaces the awkward `Schedules` button in the repository-scoped section of the left hand nav. It became extra confusing to have both a `Scheduler` and `Schedules` button.

Instead, we now have a full list of schedules inline, rendered similarly to how the pipelines and solids are rendered. A small added bonus is that you can see which schedules are running directly from the navigation menu.

The new `Scheduler` page is still an empty shell.

I'm completely open to feedback about the design, and am also happy to revert to something similar to what we had before if we don't want to prominently display the schedules like this.

Test Plan: {F226861}

Reviewers: alangenfeld, prha, bengotow

Reviewed By: bengotow

Differential Revision: https://dagster.phacility.com/D4080