Divest 'pipeline execute' command from workspace load path / being a host process
Summary:
Right now, the execute pipeline codepath is a weird hybrid between a host process and a user process - it loads the repo information via the CLI API (or grpc), but then executes it locally by loading the pipeline definition directly. This is the worst of both worlds - loading externally is slower, and executing locally gives none of the containerization benefits.

In D4078 I tried to move it to be fully a host process, but alex and nick both raised sound objections to that. as nick put it:

"just thinking about this, I feel like we have (or should have) three species of API commands:

1. Those that operate on a python module/attribute (e.g. the grpc server). These require only dagster.yaml
2. Those that operate over a workspace, and honor the contract of that workspace. These require a workspace.yaml and a dagster.yaml.
3. Those that point to a remote instance. They do not require dagster.yaml nor workspace.yaml, only a graphql endpoint.
the “execute” command feels like 1. and “launch” command feels like 2
It also makes sense in terms of internal consistency for the execute CLI API command to invoke execute_pipeline which skips all the launch infrastructure"

So this diff makes the execute pipeline command into a command of type 1 - it *only* takes in a python origin, and doesn't go through the workspace load path.

This is a breaking change, in that you're no longer to run this command by specifying a workspace. But we still have the launch command for that, and that command was already not working except for the case where your python executable matched. If we'd like to deprecate it instead and temporarily allow that codepath, I'm also open to that.

Test Plan: BK

Reviewers: schrockn, alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D4156