Generate DataFrame loader config
Separate the data for the DataFrame loader into a dict and create a
function that generates the loader config using it. This reduces
repetition and will be used to provide both the original form of the
loader config and the new form with the read key.

The descriptions of the read options are also cut down to the core
sentence instead of being a copy of the full Dask documentation.