Read dataframe using function in types
Replace the long if/else list for each read function with a lookup to
DataFrameReadTypes for metadata about the read type, which includes the
function and whether it is path based. Use this to prepare the
arguments to the specified read function, then call it and return the
resulting Dask dataframe.