Generate DataFrame materializer config
Separate the data for the DataFrame materializer into a dict and create
a function that generates the materializer config using it. This
reduces repetition and will be used to provide both the original form
of the materializer config and the new form with the read key.

The descriptions of the to options are also cut down to the core
sentence instead of being a copy of the full Dask documentation.