Write dataframe using function in types
Replace the long if/else list for each to function with a lookup to
DataFrameToTypes for metadata about the to type, which includes the
function and whether it is path based. Use this to prepare the
arguments to the specified to function, then call it.

Also, limit the AssetMaterialization yield to just file based
assets for this commit.