Patition run matrix for schedule+solid execution history
Summary:
This diff adds a matrix of partitions x runs to the bottom of the Schedule UI that makes it easier to see trends in step success / failure.

- The search bar on the left allows you to filter the UI to a particular subset of the pipeline.
- The run filter bar on the top right allows you to scope the UI to runs with a particular tag (for backfills)
- The partial red squares indicate when steps failed and then later succeeded in a subsequent run of that partition.
- The percent stats shown for each step indicate how often they transiently fail / terminally fail in partitions.
- Clicking a partition shows the individual runs beneath the matrix so you can dig deeper.

I'm not 100% sure where this UI should live - right now it's at the bottom of the schedules UI because it can become pretty long, but we could also give it a dedicated tab / page.

I've tested this with a 10x200 grid and performance is pretty good because the grid is virtualized, but we may need to do more performance tweaking as we go.

{F241419}

{F241418}

Test Plan: Run tests

Reviewers: prha, schrockn, alangenfeld

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D4090