Add dagit launch instructionsns to setup page
Summary: Somebody in #general thought that the instructions here to navigate to localhost applied to all 3 steps, not just the Dagit step. Since this is the very top of our funnel probably worth making sure it's super-clear. Clarify that it only applies to step 3 and add instructions for how execute the pipeline once you start up dagit.

Test Plan: View doc page

Reviewers: sandyryza, yuhan, alangenfeld

Reviewed By: sandyryza, alangenfeld

Differential Revision: https://dagster.phacility.com/D4309