[dagit] Handle empty repo state
Summary:
When running dagit with `--empty-workspace`, the app just shows spinners on the left nav and main content. Repair this:

- Make the repo context nullable, and begin using `loading` state to distinguish between "loading" and "empty".
- Show "No repositories found" in left nav if there are no repos.
- Render "Runs" list, but with a limited context menu to prevent re-running pipelines or opening them in the playground.
- Ensure that the logs for a single run can be displayed.

In followups, we might consider modifying the top nav to be clearer about where the tabs will go when viewing a single run. Right now, the "Runs" tab is selected for both the Runs list and the individual Run view, and tab navigation could be a little confusing when viewing the single run.

Test Plan:
`dagit -p 3333 --empty-workspace`

- Verify empty state in left nav, and that Runs are displayable in the main pane.
- View a single run, verify that logs render as expected.
- In Runs list, verify that the context menu is now limited in the no-repo state.

View Dagit in regular repo-ful state, verify that everything still looks correct.

Reviewers: alangenfeld, prha, bengotow

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D4499