[dagit] Don't crash on solid selection that isn't in subset
Summary:
Resolves #3026

When using the input to choose a solid subset, if a solid was previously selected but is not in the filtered subset, the page crashes.

Fix this by just not trying to render the selected solid.

Test Plan: View pipeline definition, choose subset, select a solid. Modify filter so that solid is no longer in the subset, verify that no errors occur.

Reviewers: bengotow, prha, sashank

Reviewed By: sashank

Differential Revision: https://dagster.phacility.com/D4714