Revert "Revert "Add retries to dagster-postgres connections""
Summary:
AKA reland D4449

This reverted under the belief that it removed connection re-use which was inaccurate.

This diff does make things faster
{P102}

at least in the grpc case - since it prevents us from duping some connections when fetching run by id

speedscope without this
{F323390}

speedscope with this
{F323401}

This also helps get things moving in the right direction for actually sharing connections so better to keep it in.

This reverts commit f9bed241cece14bbd34f47e8ba72dfff111cafbe.

Test Plan: bk

Reviewers: nate

Reviewed By: nate

Differential Revision: https://dagster.phacility.com/D4715