Improve caching in k8s example dockerfile
Summary: Often when working with the k8s-example-editable Dockerfile, we only want to update the repository code. this diff updates the image to copy over the repository code as the last step in the Dockerfile. this significantly speeds up the build, as it completely avoids the several `pip install -e` commands required for every change previously.

Test Plan: build image, change only repository code, rebuild image, verify it uses cache and builds faster

Reviewers: alangenfeld, johann, nate

Reviewed By: johann

Differential Revision: https://dagster.phacility.com/D4744