[dagit] Remove "historical" tag on Run table
Summary:
On the Run table, the "historical" tag is inaccurate for pipelines that aren't on the active repository, since we're currently just checking whether the snapshot ID matches the current snapshot ID for the pipeline name in the current repo. Let's just remove it for now.

Also "Pipeline snapshot" -> "Pipeline definition" on the Run table header.

Test Plan: View /runs, verify changes.

Reviewers: bengotow, alangenfeld, sandyryza, max

Reviewed By: sandyryza

Differential Revision: https://dagster.phacility.com/D4782