Require that start_date be on a month/day boundary for monthly/daily schedules
Summary: Right now the start date determines what time gets passed in as a partition value to the run config, which might not be what people expect (like if you start a monthly schedule on the 5th of each month, we pass in a partition with the name as the month and the value as the 5th of each month). It also makes it harder to think about timezones and ensuring that we can map from tick time to partition time, although I don't *think* there's actually any bugs here.

Test Plan: New BK coverage

Reviewers: sashank, alangenfeld, prha, schrockn

Reviewed By: sashank, prha

Differential Revision: https://dagster.phacility.com/D4787