(config-scaffold 1/n) Fix config errors visualization in Dagit
Summary: When a solid was missing an input or config in the Playground, we incorrectly displayed that the solid had valid configuration in the "error preview" section. This diff updates the error logic to display the "invalid" state when any subconfiguration for a node is missing.

Test Plan: Fill out solid config, delete one part of the solid config that was required, verify that the solid box showed an error.

Reviewers: bengotow, dish, schrockn

Reviewed By: bengotow

Differential Revision: https://dagster.phacility.com/D4807