[dagit] Quick search component (experiment)
Summary:
An initial stab at a "quick search" feature for Dagit, which could be used to quickly navigate to objects.

- Aims to ingest a GraphQL query payload for the repositories, though if we want to be able to search data outside of repos, I'd need to add that too. It may make sense for such a feature to live on the back end if we want it to do lots of heavy lifting.
- Uses fuse.js for fuzzy searching, I haven't investigated how much this adds to the built JS but I didn't want to try rolling something like this myself.
- Currently searches on pipelines, plus the schedules, runs, and assets that apply to those pipelines.

Feedback very welcome. If this is something that doesn't seem like it would be useful in Dagit, by all means say so!

Video inline.

Test Plan: Run storybook, use SearchDialog story to perform searches and stuff.

Reviewers: bengotow, alangenfeld, catherinewu, prha, dgibson

Reviewed By: bengotow

Differential Revision: https://dagster.phacility.com/D4876