Differentiate between ModuleCodePoitner and PackageCodePointer on the gRPC server
Summary:
Working on a refactor of the loading path and noticed that we're currently assuming that the code pointer is a ModuleCodePointer and never a PackageCodePointer. This will break schedule origins when people switch from cli to grpc if they were using a PackageCodePointer before.

This diff adds code to ensure that the code pointers are the same when moving from cli to grpc.

Separately, we should really make python_package the default and kill ModuleCodePointer in 0.10.0.

Test Plan: New and existing BK

Reviewers: prha, alangenfeld, sashank

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D4911