Update SqlRunStorage runs query to use inner join
Summary:
We don't need to do an outer join between the runs table and the runs tags table when querying the run storage with a tags filter.

This diff updates the join to be an inner join. Credit to @sandyryza for the find.

Test Plan: Existing tests in `test_fetch_by_filter`

Reviewers: sandyryza, alangenfeld

Reviewed By: sandyryza

Subscribers: sandyryza

Differential Revision: https://dagster.phacility.com/D4928