Change asset key storage structure, make asset key strings more permissive
Summary:
Changes the asset key storage to be an JSON array instead of a custom
delimited string representation.  Relaxes the AssetKey parameters to be
permissive instead of raising an Exception.

Could also consider a schema migration and supporting both a structured json
field as well as the legacy delimited field, so we don't have to do these OR sql
clauses (and deduping).  That has a downside of forcing a migration, and so would
need to delay this until the 0.10.0 launch.

Test Plan: bk

Reviewers: schrockn, dgibson, sandyryza

Reviewed By: sandyryza

Subscribers: sandyryza

Differential Revision: https://dagster.phacility.com/D4944