[dagit] Empty definition state for empty pipelines
Summary:
Resolves #3094

We incorrectly show the "large DAG" notice on empty pipelines in the Definition view. Show an empty state instead.

Test Plan: Create an empty pipeline, load it in dagit. Verify that the empty state appears. Load a large DAG, verify that it continues to show the warning. Load a normal pipeline, verify normal behavior.

Reviewers: sashank, prha, alangenfeld

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D4969