Remove `download-mocks` script
Summary:
It seems like maybe SVG mocks have been broken for quite a while, and we should be able to replace this behavior with Apollo mocking.

With my change to use `react-scripts test`, I deleted a bunch of jest config, which breaks the existing `download-mocks` script further.

So in the meantime, delete the `download-mocks` script so that no one tries to use it.

Test Plan: Buildkite

Reviewers: bengotow, prha, yuhan

Reviewed By: yuhan

Differential Revision: https://dagster.phacility.com/D4983