[dagit] Gracefully handle repo location failures on start
Summary:
Resolves #2918

Create a repository location error recovery pathway, for loading errors on initial Dagit startup as well as reload errors.

- Display a clear warning on the left nav that indicates that an error has occurred.
- Link to a page with all loaded repositories and tags indicating their loaded/failed status.
- Provide a link to view the error message in a dialog.
- After the issue is resolved, the user can click "Reload" in the table view, or use the reload button.
- The GraphQL query is refetched, and the repository locations are updated accordingly.

Video in a comment below.

Test Plan: Load workspaces that have no repos, broken repos, or only working repos. Verify that the error appears on Dagit startup for broken repos, and that I can use the "View error" and "Reload" options as described in the summary.

Reviewers: dgibson, alangenfeld, bengotow

Reviewed By: dgibson

Differential Revision: https://dagster.phacility.com/D5034