fix fan-in input mappings in composites
Summary:
Currently if you attempt to fan-in any input mappings in a composite definition we error out. This takes an approach of fixing this without changing the input/output mapping APIs. It's not obvious to me how the output mappings would work in that version of the world, and breaking changes have their own inherent cost. This approach is a little awkward, but I think maybe the way to go.

depends on D5217

resolves https://github.com/dagster-io/dagster/issues/3176

Test Plan:
added test cases

manual verify snapshots including fan-in input mappings render properly in dagit by executing the composition toy pipeline and loading the snapshot in `dagit --empty-workspace`

Reviewers: schrockn, sandyryza

Reviewed By: schrockn, sandyryza

Differential Revision: https://dagster.phacility.com/D5268