Allow config yaml dump to include unicode
Summary: When resolving run config yaml, we currently dump to strings that don't allow Unicode, resulting in non-ASCII characters rendered as code points. This leads to some confusing config output for anyone that uses Unicode in their configs. Allow unicode in the yaml dump to resolve this.

Test Plan: Perform a run with "Москва" in the config. Open config from runs table view, verify that the config display shows the actual Unicode characters instead of code points.

Reviewers: alangenfeld, prha, schrockn

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D5182