[dagit] Allow Run execution for any pipeline in the workspace
Summary:
Resolves #3254

Allow run re-execution from any location in Dagit, regardless of what the left nav repo switcher says. We do this by finding the repository that matches the specified `PipelineRun`, either by using its `RepositoryOrigin` or, in the case of older runs, by looking for a pipeline name match.

This fixes the issue in which a run cannot be kicked off if the "active" repo in the left nav has not been switched to the Run's pipeline's repo.

Test Plan: View various places in Dagit where I can kick off runs: Runs page, a previous Run page, Playground. Verify that I can execute runs for any pipeline that exists in the workspace repositories, regardless of what the left nav says.

Reviewers: dgibson, prha, johann, alangenfeld

Reviewed By: dgibson

Differential Revision: https://dagster.phacility.com/D5251