Mock the rest of AWS in tests (#3267)
With this change, all dagster-aws tests (except for those hiding behind environment variable flags that require live resources) can run locally using mocked resources. This means #2302 can be resolved and hopefully it will make it easier for contributors to extend dagster-aws.

I didn't touch any implementation code. I rewrote a few tests to be a little simpler but otherwise, this is a pure refactor of test code to use the mock fixtures instead of the combination of real resources, magic mocks, repeated uses of the @mock_s3 decorator that they were already using. I do think there remains a lot of opportunity to improve the tests - as well as some of the implementations - of a lot of dagster-aws to make them more consistent, simple, and readable but those are all tasks for another day 😉