Get Athena results
Dagster's various SQL-ish query interfaces expose a variety of different
ways to return results:

- dagster_aws.redshift.redshift_resource.execute_query returns a list
of tuples
- dagster_gcp.bq_solid_for_queries returns a Pandas
dataframe of results
- dagster_snowflake.snowflake_resource.execute_query() returns a list of
tuples

This follows the patterns established by the Redshift and Snowflake
implementations to return a list of tuples.

There are two ways to get query results from an Athena query:

1. Call GetQueryResults
2. Every SELECT Athena query saves its results as a CSV file in S3

While the former is probably the more "correct" way of retrieving
results, the latter is far simpler to implement and mock out for
testing. A consequence of the choice to use the CSV file is that all
columns are cast to strings.

Like the Redshift and Snowflake resources, retrieving results is
entirely optional (and I'd probably recommended against it unless the
expected dataset is relatively small).