Example docker deployment with daemon and user code deployment
Summary:
This provides a sample docker-compose setup for deploying dagster in docker (without k8s) that we could reference in the docs as a sample deployment. There's a container for dagit, the dagster-daemon, and user code, and each run is also launched in its own container using the DockerRunLauncher introduced in D5348.

Once we have 0.10.0 available in pypi, the Dockerfiles (especially dagit) can be made simpler since they no longer need to build from source.

Still need to figure out the best way to test this, would welcome some initial feedback in the meantime though.

Test Plan: TBD

Reviewers: nate, johann, max, alangenfeld, catherinewu

Reviewed By: alangenfeld

Subscribers: sashank, schrockn, prha

Differential Revision: https://dagster.phacility.com/D5416