Checkpoint - why can't my dagit connect to the postgres DB???
Summary: Just seeing if the solution here jumps out to anybody else, this is extremely WIP. I have a docker-compose with a postgres node and a dagit node, and the latter is unable to connect to the postgres DB on startup. I think there's something obvious about docker networking and/or setting up the postgres DB that I must be missing.

Test Plan: WIP

Reviewers: nate, johann, max, alangenfeld

Differential Revision: https://dagster.phacility.com/D5397