[dagit] Avoid flashing progress bar in Loading
Summary:
The `Loading` component currently shows a progress bar during the loading state of the supplied `queryData`. This often means a brief flash of the progress bar, then the fully rendered child of the `Loading` component. This can be jarring, and is often unnecessary.

Instead, wait a brief moment during loading before showing the progress bar. If the loading completes during that wait, the contents render without the progress bar flash. If the loading does not complete during the wait, show the progress bar as normal while the loading finishes.

Test Plan: View a bunch of pages on Dagit that use `Loading`. Verify that when loading is completed quickly, the progress bar does not flash.

Reviewers: bengotow, sashank, prha

Reviewed By: sashank

Differential Revision: https://dagster.phacility.com/D5410