output manager
Summary:
Introduces OutputManagers.  An output manager is like an AssetStore, but without a get_asset function.  I.e. it takes responsibility for materializing the output of a solid, but doesn't take responsibility for loading it in downstream steps.

OutputManagers are configurable per output.  For any run where any output on a solid has a configurable OutputManager, none of that solid's outputs can be materialized with a type materializer.

In this diff, I have studiously avoided any consolidation of AssetStores with OutputManagers. That consolidation will happen in a followup, in which AssetStores will be renamed to ObjectManagers and extend OutputManagers.

Depends on D5408.

Test Plan: bk

Reviewers: yuhan, schrockn, alangenfeld

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D5427