Make the run returned by DagsterInstance.submit_run be already in PipelineRunStatus.QUEUED
Summary:
This is useful for tests and graphql callsites - ensures that the run that is returned in the graphql mutation response has the correct state.

We should probably do something similar with the default run coordinator and/or run launchers and have it return STARTING once that is a state.

Test Plan: BK

Reviewers: johann, prha, alangenfeld

Reviewed By: johann

Differential Revision: https://dagster.phacility.com/D5431