[Addressing Removal 1/2] Fixed display CLI to work with asset store.
Summary:
The recent deprecation of memoized execution on intermediate storage broke the memoized display CLI. This diff updates the CLI to use the asset store instead of rifling through addresses.

An unfortunate result of this change is that the asset store API doesn't provide a convenient way to surface where results are stored. Until this is figured out, I've changed it to a description of what will happen with the output in the "next" run.

Test Plan: Memoized display cli tests

Reviewers: sandyryza, yuhan, alangenfeld

Reviewed By: sandyryza

Differential Revision: https://dagster.phacility.com/D5435