[dagit] Timezone display with user timezone info
Summary: Move `TimezoneDisplay` into its own component and use it in ScheduleDetails and SensorDetails, with a change to add a "clock" icon that shows the time in the user's own timezone as needed.

Test Plan: View Schedule and Sensor pages, verify that the timestamps appear with the timezone tooltip icon if the displayed time has a different timezone from mine (CST).

Reviewers: dgibson, prha, johann

Reviewed By: dgibson

Differential Revision: https://dagster.phacility.com/D5463