[dagit] Handle more error flavors on failed backfill
Summary: Backfill errors apparently come in many different varieties. Handle more of them so that we can surface useful error messages instead of the empty dialog, which is what it currently shows.

Test Plan: Launch a backfill, see error. Click to view the error, verify that it now shows me something of value.

Reviewers: bengotow, alangenfeld, cdecarolis

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D5616