[dagit] Include previous year for default timestamp format
Summary:
In our default timestamp format, we don't currently display the year unless the timezone is UTC. This means that in January, we'll be showing 2020 timestamps without the year. This is probably not a problem right now, but I expect that as users are using Dagster longer, they will start to see older and older data, and having a year on their timestamps will become more important.

With this change, when determining the default timestmap format, we check whether the user's current year (based on their timezone setting) is the same as the year for the supplied timestamp. If it is not the same year, display the year in the formatted string.

Test Plan: Force the year to be different on run timestamps on the Runs page. Verify that years appear.

Reviewers: bengotow, prha, dgibson, johann

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D5645