[dagit] Improve messaging on backfill submission dialog
Summary:
Resolves #3358

In the backfill creation dialog, instead of saying that we are "launching" runs, say that we are submitting them. That way it is hopefully a bit clearer that the user doesn't need think they need to wait for all runs to complete.

Update the run creation button generally to use "Submitting...", and block the backfill dialog from being closed while the submission is in process.

Test Plan:
View backfill creation dialog, submit some runs. Verify new text, including when steps are selected. Verify that I can't close the dialog while the runs are being submitted.

View single run, re-execute. Verify that the button label has changed there as well.

Reviewers: sandyryza, dgibson, prha, johann

Reviewed By: sandyryza, dgibson

Differential Revision: https://dagster.phacility.com/D5719