Warn on multiple daemons
Summary:
Use the daemon_id field, and warn when it changes. Downside is that we'd be doing an additional read every time we write a heartbeat. Optionally could add smarts so that there is no warning due to an old heartbeat.

Would it be more useful if the daemon_id was a process ID or something like that and was surfaced in the heartbeat UI? The issue I see is that a relevant ID depends on the deployment, e.g. on K8s you'd use hostname to get the name of the pod but in other deployments that's not unique.

Test Plan: existing, tried running multiple daemons and saw warnings

Reviewers: dgibson, prha

Reviewed By: dgibson

Differential Revision: https://dagster.phacility.com/D5835