Run ECS tasks
This replaces the existing ECSClient with new implementation. For now,
it only implements run_task.

You can use this client to run an already defined Task Definition on an
existing cluster using the FARGATE launch type. Task counts are
hardcoded to 1.

Moto doesn't yet support the FARGATE launch type (nor does it support
status transitions) so most ECS calls are stubbed out similarly to the
dagster-aws Athena implementation.

ECSClient.run_task implements its own polling instead of using the
TaskStopped Waiter:

https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/ecs.html#ECS.Waiter.TasksStopped

This is because I couldn't get the Waiter to play nicely with botocore's
Stubber; it passes a "waiterConfig" param to the DescribeTasks endpoint
but the Stubber doesn't accept that as a valid param.