[dagit] Don't disappear Daemon statuses when loading
Summary:
Resolves #3517

Since we're using `network-only` and hiding Daemon statuses when the query data isn't available yet, the table disappears and reappears a little too often.

@sashank, is there any cause for concern with changing this query to `cache-and-network`? I don't remember if there's anything related to your GRPC changes.

Instead, use `cache-and-network` to hang on to data that we have while refetching, and render a spinner when loading is in progress.

Test Plan: View `/instance`, use repository location reload and the countdown refresh. Verify rendering and behavior described above.

Reviewers: sashank

Reviewed By: sashank

Subscribers: sashank

Differential Revision: https://dagster.phacility.com/D5888