[dagit] RFC: Instance status dot in left nav
Summary:
Add a "status dot" to the Instance status left nav item in the updated nav. Worth discussing a bit to figure out how to make this appropriately useful.

- It would be good to make the repository error display less prominent in the new navigation. (https://github.com/dagster-io/dagster/issues/3490) The dot will ideally provide a summary of error/warning state that is noticeable but also easy to disregard if the user doesn't need to care about it.
- When the status is still loading, show it as "gray".
- Currently, "green" means that there are no repo errors and that all daemons are healthy. Anything else is "yellow".
  - More ideally, we would know whether Dagit is being run in a dev or prod mode, so we could be clearer about what's truly an error state. For instance, in prod, it may be that any daemon error should be surfaced as a "red" state, but this is a "yellow" state in dev. We don't have such a flag right now.
- I renamed the left nav item from "Status" to "Instance". I don't know if this is the right thing to do.

I'm thinking it might also be useful to show an `Alert` on the top of the "Instance status" page to give a bit more context about /why/ the status dot is what it is when first coming to the page.

Test Plan: Toggle left nav changes on. Verify status dot behavior.

Reviewers: bengotow, alangenfeld, dgibson, prha

Reviewed By: bengotow

Differential Revision: https://dagster.phacility.com/D6403