create run partition data migration, tie to `dagster instance migrate`
Summary:
Sets up migrations for run_storage, to be run upon `dagster instance migrate`.

Currently, we have a CLI command `dagster instance migrate` which performs schema migrations, and `dagster instance reindex` which performs optional data migrations (over the event log).

https://dagster.phacility.com/D5840 adds a partition column to the runs table, which would require a data migration in order to begin doing reads off of that column.

This diff sets up the machinery to track that migration status, and also notably performs a data migration over the runs table upon `dagster instance migrate`.

The upside is that with `0.10.0`, we have some schema changes that must be performed and so we know that people will run `dagster instance migrate`.  The downside is that this is definitely a data migration and its latency depends on the size of the runs table.

Test Plan: bk

Reviewers: dgibson, alangenfeld, schrockn

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D5916