Default user deployment replica count to 1
Summary:
Resolves https://github.com/dagster-io/dagster/issues/3546.

Since `userDeployments.deployments` is a list,
when it is overriden, the helm defaults are
not persisted. As a result, any overrides must
fully meet the json schema.

Since `replicaCount` usually defaults to 1 on
most cases, we make it optional, and set the
default during templating.

Test Plan: N/A

Reviewers: catherinewu, nate, bob

Reviewed By: catherinewu

Differential Revision: https://dagster.phacility.com/D6039