[Memoized Dev Example 1/2]Add Memoizable IO Manager to API Docs, fix typo in Memoizable IO Manager documentation.
Summary:
This diff adds the Memoizable IO Manager to the API docs. This is necessary for linking back to the example.

I chose to put the Memoizable IO Manager under the internals section, not the IO Managers section, since it isn't exposed in the public-facing API yet.

Test Plan:
{F511541}
Test docs build

Reviewers: sandyryza, yuhan

Reviewed By: sandyryza

Differential Revision: https://dagster.phacility.com/D6046