change some names in execute_step output handling
Summary:
This code has seen a lot of churn lately and some of it has gotten a little stale.

I think the "event sequence" naming is a little cumbersome.  I had trouble figuring out how to phrase some of the names in that language.  Now that we have type annotations, it feels a little less necessary.  Can find a way to make it work if others feel strongly though.

Test Plan: bk

Reviewers: alangenfeld, yuhan

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D6162