[dagit] Set per-snapshot run filters as "permanent"
Summary:
Resolves #3598.

There's some weird behavior right now when adding or removing filter tags on the Runs view for a snapshot or pipeline, which seems to be related to using the URL querystring to track filter state. See attached GitHub issue for an example of the bug.

To resolve this, pull the pipeline and/or snapshot filter out of the querystring state and filter input, and render them as their own static tags. They aren't supposed to be removable from the input at all, and we don't need to track them in the querystring (since we're already tracking them in the URL path) so we may as well just show them as static.

Test Plan: View Runs tab for snapshot and for pipeline. Verify that these "static" tags are rendered as such, and that filtering works correctly.

Reviewers: bengotow, prha, dgibson, johann

Reviewed By: bengotow

Differential Revision: https://dagster.phacility.com/D6256