Force scrollbar visibility on partition matrix to make behavior more clear #3638
Summary: The partition matrix scrolls horizontally but some users on macOS are finding this confusing because the scrollbars hide when not in use. To make it extra clear, we can use CSS to make a visible scrollbar that doesn't go away. (Note: on Firefox this is not possible and the scrollbar still disappears until you hover).

Test Plan: Run tests

Reviewers: dish

Reviewed By: dish

Differential Revision: https://dagster.phacility.com/D6490