[dagit] Make repo switcher more flexible
Summary:
Moving in the direction of allowing the selection of more than one repo in left nav filtering, make the current repo switcher and left nav display more flexible.

- Stop switching repository context in the left nav based on navigation in Dagit. When the user sets repo context, leave it alone until they change it.
- Change all pipeline/solid/schedule/sensor lists to allow showing results for an array of repos, not just one.

This requires doing a bit of refactoring of how we handle "workspace context", in that there is no longer an "active" repo to query with. This introduces some complexity in some of our lazy left nav queries, e.g. solids.

The next step here is to replace the existing repo switcher with one that allows selecting multiple repos in the left nav, which will be in a followup.

Test Plan: View Dagit, verify that the left nav repo context remains the same no matter where I navigate in Dagit. Switch repos, verify that all objects are fetched and rendered correctly in the nav.

Reviewers: bengotow, sandyryza, prha, sashank

Reviewed By: bengotow

Differential Revision: https://dagster.phacility.com/D6893