[dagster-postgres] Cleanup Event Log
Summary:
There was a bit of dead code in the Event Log Listen/Notify implementation & the lack of type annotations made some of the logic hard to follow. To fix this, removed the dead code & added type annotations where appropriate.

Also there was (fixed with this diff) a small bug in `pynotify` in the processing of `Notify` objects - if multiple are received at once, we were processing the notifications in reverse (e.g. if we received events 3 & 4 simultaneously then we were formerly processing 4 and then 3 - this lead to some flakyness for some of the PG test suite).

Test Plan: bk

Reviewers: rexledesma, max, prha, bob

Reviewed By: max

Differential Revision: https://dagster.phacility.com/D6862