Don't pass instance to launch_run (#3551)
* Don't pass instance to launch_run

The `RunLauncher` abstract base class requires implementing classes to
pass `instance` to `.launch_run()`. Few of them do anything with the
argument (other than the occasional type check) because they've mostly
been updated to instead reference `._instance` that gets set during
`RunLauncher.initialize()`.

This leads to a number of confusing tests where `None` is passed for the
`instance`. I've started implementing a new `RunLauncher` for ECS and I
think it makes sense to simplify the interface rather than to introduce
another place where the argument is passed but unused.

The only `RunLauncher` with a functional change in this commit should be
to `DockerRunLauncher`; previously it was the only one still using
`instance` but now it has been updated to instead use `._instance`

* Call instance method

* Remove unused import