Add check-manifest tests for package contents
Summary:
This adds a test to our buildkite suite called 'manifest' that installs `check-manifest` and runs it against every library package.

This diff removes some empty `requirements.txt` files, but otherwise does not change the package contents in `sdist`.

Instead it configures `setup.cfg` to instruct `check-manifest` to ignore certain files that are not included.  We should potentially reconsider and start including test files and examples in the manifest, but that should probably be a separate diff.

This would have caught the manifest bug fixed by 4d1cec096.

Test Plan: bk

Reviewers: dgibson, alangenfeld

Reviewed By: dgibson

Subscribers: rexledesma

Differential Revision: https://dagster.phacility.com/D7116