change DagsterInstance.get behavior
Summary:
Our setup has matured and changed a lot since this pattern was first introduced.

The main change here is to  change `DagsterInstance.get` to only return a `DAGSTER_HOME` based instance.

To maintain the gracefully / gradual getting started with dagster path, we still allow you to use `dagit` without it set. The change here is to be explicit that we are using a temporary directory and that DAGSTER_HOME can be used for persistence.

resolves https://github.com/dagster-io/dagster/issues/3717
resolves https://github.com/dagster-io/dagster/issues/3817

Test Plan: bk

Reviewers: max, dgibson, sashank, prha, sandyryza

Reviewed By: max, sandyryza

Differential Revision: https://dagster.phacility.com/D7261