[dagit] Backfill dialog: separate names from statuses
Summary:
Resolves #3783

Separate the retrieval of partition set statuses from the retrieval of the partition set names themselves. Ideally this should speed up initial load of the backfill creation dialog, with the statuses only retrieved when the user clicks "Re-execute from failures".

Test Plan: Open "Launch backfill" dialog from a partition set. Click "Re-execute from failures" checkbox, verify that the statuses are queried and loaded, and that the grid is disabled and a spinner appears during the loading state. After loading is complete, verify that selection behavior is correct, that I can select failures in the grid.

Reviewers: bengotow, sandyryza

Reviewed By: sandyryza

Differential Revision: https://dagster.phacility.com/D7338