download csv in tutorial
Summary:
Preview: https://dagster-git-download-csv-elementl.vercel.app

"When should I use a solid and when should I use a regular Python function?" is a question that came up a couple times last week. I believe the examples in our current tutorial give a misleading answer to this question, which makes it difficult for prospective users to reason about the value of Dagster.

This diff posits that the central reason why it's useful to split code across multiple solids is so that it's easy to re-run some code without re-running everything.

In light of that, it replaces the solid that reads a CSV with a solid that downloads a CSV - it's useful to be able to re-run downstream steps without re-downloading the CSV.

This has some small side benefits as well:
* Removes a tutorial setup step.
* Removes reading from `__file__`, which might throw some readers off

Test Plan: manual inspection, bk

Reviewers: owen, max, yuhan, sashank

Reviewed By: owen

Differential Revision: https://dagster.phacility.com/D7516