ensure that sets and frozensets are always serialized in the same order inserdes
Summary: frozensets and sets seem to choose their order randomly. This can cause the same pipeline to produce a different snapshot non-deterministically. This diff makes it so that sets and frozensets are serialized and deserialized in a consistent order, so that they will always produce the same snapshot and snapshot ID.

Test Plan: this is very tricky to test specifically since the order appears to be non-deterministic but often the same in a single process (in particular making it change between serialization and deserializaiton is very difficult). but if you try enough times with a daemon in a certain environment and dagit in a certain other environment it is very easy to trigger this (and it now longer triggers)

Reviewers: alangenfeld, max

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D7518