[dagit] Link to pipeline from Run table
Summary:
A modest proposal to link from the Runs table to pipelines: show a linked icon next to the pipeline name.

The issue is that when we link the pipeline name, it looks like the "primary" link in the row, and users tend to click that link when they're trying to navigate to the run. That's why the pipeline name is currently not linked at all.

The idea here is that this is a small link that is hopefully clear in purpose, but is unlikely to be accidentally clicked when the user wants to view the run.

Test Plan: View Runs page in Dagit, verify rendering and behavior of pipeline link icon.

Reviewers: bengotow, schrockn

Reviewed By: bengotow

Differential Revision: https://dagster.phacility.com/D7572