[dagit] Fix favicon for path-prefix case
Summary: The favicon path is broken for the path-prefix case. Fix this by restoring the `link` tag in index.html so that it can be modified by app.py, and update the custom run-status favicon paths using the `basePath` value.

Test Plan:
`yarn build-for-python`, run dagit with and without path-prefix. Verify that favicons load correctly, and update when navigating to runs.

Run in dev mode too, verify that everything works.

Reviewers: bengotow, prha

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D7567