Extract should_cleanup fixture
Summary:
By treating this as a fixture, we provide it to other fixtures without directly passing `pytestconfig` around as an argument.

By breaking it into its own fixture instead of keeping it in `_helm_namespace_helper`, we can simplify the responsibility of an already fairly complicated function.

Depends on D7581

Test Plan:
unit
integration

Reviewers: rexledesma, johann, dgibson

Reviewed By: rexledesma

Differential Revision: https://dagster.phacility.com/D7582