[dagster-dbt] re-enable asset materializations when running `dbt cli` solids
Summary:
Fix for https://github.com/dagster-io/dagster/issues/3815.

The schema of `run_results.json` has changed as of DBT 0.19.0. This breaks the asset materializations in our dbt solids. This diff re-aligns the way we construct materializations such that materializations are yielded again.

Test Plan: Add to `test_dbt_cli_run` to make sure that the correct asset materializations are yielded when using `run` CLI (this was previously not covered under test). Tested fxnality on both 0.19.x and 0.18.x

Reviewers: sandyryza, max

Reviewed By: sandyryza

Differential Revision: https://dagster.phacility.com/D7832