Add more retries for Helm tests
Summary:
Because these tests call to the $ref's of the JSON Schema,
they are subject to failure. Bump up the retry number on
Buildkite.

Also only run the tests on python 3.8. Running on other
versions of Python is repetitive and is a source of flakiness
Helm tests are just template tests is not an importable
library in Dagster.

Test Plan: bk

Reviewers: alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D7935