[schedule/sensor testing 1/n] [RFC] build_schedule_context
Summary: (Depends on https://dagster.phacility.com/D7940) This diff provides a convenient fxn for constructing a schedule context. Right now, it is just a function, but there is an argument for making it a context manager that defaults the instance to `instance.get()`.

Test Plan:
Convert unit test decorators to use build_schedule_context instead of directly constructing ScheduleExecutionContext

API reference:
{F820894}

Reviewers: sandyryza, alangenfeld, prha

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D7942