[examples] Adding a Run Attribution Example via a Custom Run Coordinator
Summary:
As part of an ask by a user, this is an example of how one would perform Run Attribution by reading off of HTTP headers in the Run Coordinator.
Depends on D8030, D8187

Test Plan: Header parsing logic was tested both manually (by replacing the `jwt_token` in test with the one provided by the user) and by creating a similar `X-Amzn-Oidc-Data` header to the one used by the user.

Reviewers: catherinewu, rexledesma, jordansanders

Reviewed By: rexledesma

Differential Revision: https://dagster.phacility.com/D8043