[mypy] logger definition
Summary:
Type annotations for logger definition in preparation for build_logger_context.

It's a little weird that I needed to add a `__call__` here, because we don't do the `update_wrapper` stuff we do elsewhere. Instead, it seems mypy is unable to parse that certain loggers are invoking the decorator s.t. they produce a logger definition vs a function that produces a logger definition.

Test Plan: mypy

Reviewers: alangenfeld, sandyryza

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D8063