TestStepHandler wait for processes to exit
Summary: This test formerly exited once a STEP_SUCCESS event was emitted for every step. However there is a gap between when a step worker emits that event, and when the step worker's process exits. This caused flakes on windows, where processes cannot still be alive when a test exits.

Test Plan: existing test

Reviewers: dgibson, max

Reviewed By: dgibson

Differential Revision: https://dagster.phacility.com/D8081