[dagit] Don't show Asset alert if `asOf` points to latest
Summary:
When loading an asset materialization view, load the most recent materialization so that we can compare it to the `asOf` value, if any.

If there is an `asOf` value and the `asOf` materialization and most recent materializations have unequal timestamps, show the `Alert`. Otherwise, we're looking at the most recent materialization and do not need to show the `Alert`.

There is an unrelated issue in which `asOf` is not inclusive, which will be fixed separately.

Test Plan:
View asset with:

- `asOf` parameter that is known to be prior to the latest materialization. Verify `Alert`.
- `asOf` parameter that is in the future. Verify no `Alert`.
- No `asOf` parameter. Verify no `Alert`.

Reviewers: bengotow, prha, sandyryza

Reviewed By: prha

Subscribers: schrockn

Differential Revision: https://dagster.phacility.com/D8084