Revert tutorial changes for solid invocation, make testing docs around solid invocation experimental
Summary: Solid invocation is still undergoing active development / changes, so we should keep it experimental for now. As a result, we have to make sure not to introduce a docs regression. Rolling back tutorial updates for now until invocation is in a more stable state.

Test Plan:
unit tests + test build:
{F848662}
{F848667}

{F848666}

Reviewers: sandyryza, yuhan

Reviewed By: yuhan

Differential Revision: https://dagster.phacility.com/D8147