[dagit] Add a polling query for Run logs
Summary:
Branch `LogsProvider` to use a polling query in cases where the websocket is not open. The query uses `after` to retrieve slices of logs.

Depends on https://dagster.phacility.com/D8220.

Test Plan: Force query branch in code. Launch a run. Verify that polling occurs correctly, and that the Gantt and logs table update correctly as well. Verify that the polling terminates properly when the `status` reaches a completed state.

Reviewers: alangenfeld, dgibson, bengotow

Reviewed By: bengotow

Differential Revision: https://dagster.phacility.com/D8224