Merge Workspace into WorkspaceProcessContext, Make WorkspaceRequestContext an IWorkspace
Summary:
Right now in master there is Workspace, WorkspaceProcessContext, and WorkspaceRequestContext. You can fetch repository locations from both Workspace and WorkspaceRequestContext (although its not thread-safe to do so from Workspace).

The goal of this diff is to consolidate those concepts a bit so that we can start passing around workspaces into things like the run launcher in a way that is not extremly confusing. After this diff, the thing that we currently call WorkspaceRequestContext is the actual 'workspace' (since it's the thing that it's safe to get repo locations from) and the WorkspaceProcessContext is the process-scoped object that knows how to generate workspaces on each request. Then we can pass in either a WorkspaceRequestContext (what dagit uses) or a DynamicWorkspace (what the daemon uses) in when you need an IWorkspace for something to fetch repo locations from.

Open to name suggestions that make this distinction more clear - e.g. could call WorkspaceRequestContext FlaskRequestWorkpace or something.

Test Plan: BK

Reviewers: alangenfeld, sashank

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D8561