Merge explorer into the Overview tab, remove Definition tab for all users
Summary:
This diff consolidates the "Definition" and "Overview" pipeline tabs for all users (feature flagged or otherwise). The pipeline's DAG is shown on the left, and the right sidebar shows an expanded "Info" tab with sections for everything on the old overview tab.  The info tab is more or less the same with and without the feature flag. The only differences are that with `pipeline:mode` tuples, the schedules, sensors, and modes displayed are only the ones bound to the mode, and references to "pipeline" are swapped out for "job".

--

I used a context to inject a separate version of the Info tab into the sidebar when the pipeline explorer is used for the Overview tab as opposed to the Graph explorer pages. I set up a context rather than using props.children or passing the data down into the tree because I think it'd be nice to sprinkle more job- or run- related stuff into the DAG as well, and we might pass down in a single run's logs to colorize the DAG, etc. and need data from the top container all the way down in the tree. If it doesn't pan out I think we could change it.

This diff also fixes the scrolling issue caused by styles on the graph pages and cleans up the pipeline explorer so it doesn't internally modify the explorerPath and can be mounted more flexibly.

Misc:

- The `/overview` URL path is gone. For our existing set of "explorerPath" helpers to work, the URL has to contain `/pipeline:mode~query/solid/solid` and I didn't want to mess with it in this diff. 🙈

- The explorer sidebar was choosing tabs in a super strange way (my fault i'm sure) and I changed it to use a simple `?tab=types` or `?tab=info` query string for tab state + updated places where this was set.

{F1052410}

Test Plan: Give it a spin, run snapshots

Reviewers: dish, sandyryza, alangenfeld, schrockn

Reviewed By: dish

Differential Revision: https://dagster.phacility.com/D8614