[dagit] Better usage of space in RepoNavItem
Summary:
Now that we have a little more horizontal space in the left nav, let's make some better use of it for the repo nav item.

- Occupy a higher max-width to show as much of the repo name as possible
- Change DOM structure so that the "Filter" button can sit above the refresh button, if the repo name is long
- Add a title to the repo name so that the full string shows up on hover

Test Plan: Verify that basic rendering is correct with single repo or multiple repos selected. Use dev tools to make the repo name longer, verify that it truncates nicely.

Reviewers: bengotow, prha, yuhan

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D8719