Give dagit permission to interact with ECS
Summary:
Two different users encountered suspiciously similar issues with
ECS depoyments today:

https://github.com/dagster-io/dagster/issues/4377

and

https://dagster.slack.com/archives/C01U954MEER/p1626814414314800

I was able to reproduce the issue by launching a run and then
immediately switching to view in progress runs. I started seeing GraphQL
failures indicating NoCredentialsError. I suspect this is because in our
reference ECS deployment, dagit indeed doesn't have the ability to
interact with the ECS SDK.

This adds a TaskRole to dagit with the same widely-scoped access as the
one we attach to the daemon task. We should more narrowly scope both but
I'll leave that task to another day.

Test Plan:
I tested this by redeploying using the updated docker-compose.yml and
verifying that:

a) the TaskRole is attached
b) I can't reproduce the error

Reviewers: dgibson, johann, max

Reviewed By: dgibson

Differential Revision: https://dagster.phacility.com/D8977