Infer cluster from tags
Summary:
In addition to the credentials error exposed in
https://dagster.phacility.com/D8977, I noticed that we don't actually
need to query `self._task_metadata` to get the cluster name when
terminating. This is because we tag ECS launched pipeline runs with both
the ARN and the cluster. If it's lacking either, we can't terminate it
regardless of what `self._task_metadata` says.

Test Plan: unit

Reviewers: dgibson, johann, max

Reviewed By: dgibson

Differential Revision: https://dagster.phacility.com/D8978