feat(helm): set compute log manager default as noop
Summary:
In a Helm deployment of Dagster, it does not make sense to default to the
local compute log manager as pipeline runs are executed in ephemeral jobs.
In a Kubernetes environment, users should default to using a cloud provider
to host their compute logs.

We should deprecate the local compute log manager as configuration in our Helm
chart. Now, we default to the noop compute log manager. To prevent this from being
a breaking change, if the user previous specified the local compute log manager,
it will now noop (since it wasn't doing anything in the first place).

Test Plan: pytest

Reviewers: alangenfeld, prha, johann

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D9038