[dagit] Modify App initialization config
Summary:
A couple things:

- The `subscriptionParams` config value is not currently used
- We want to have the flexibility to add headers to `HttpLink`

These headers will also be applied as the `connectionParams`. This is a little awkward, but seems like it might be the best way for us to get initialization info to the server.

Test Plan: Buildkite. Run Dagit, verify that all http and ws connections behave correctly.

Reviewers: dgibson, bengotow, prha

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D9065