[dagit] Fix Gantt chart time indicators offset from step edges when scrollbars present #4408
Summary: The ticks sit on top of the Gantt scroll view and not inside it, so they need to be positioned accurately based on the size of the scroll view beneath them. The viewport measurement helper was using getBoundingCLientRect() which includes the scrollbars in the width + height of the element and we needed to use clientWidth / clientHeight instead.

Test Plan: Observe problem is fixed with mouse plugged in

Reviewers: dish

Reviewed By: dish

Differential Revision: https://dagster.phacility.com/D9102