Retry ECS describe_tasks when launching
While implementing https://github.com/dagster-io/dagster/pull/4428 I
noticed that there's actually one more place where we describe tasks -
when gathering ECS metadata. This one is a bit more tricky than all of
the termination related requests because we don't want to fail to launch
a task immediately if ECS gives us an inconsistent response.

Instead, I'm following the ECS recommendation to exponentially backoff:

https://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_RunTask.html

> Apply an exponential backoff algorithm starting with a couple of
> seconds of wait time, and increase gradually up to about five minutes
> of wait time.

However, I'm _not_ waiting the entire 5 minutes in this initial
implementation. Instead, I'm waiting ~1 minute (51.1 seconds to be
precise since that's how long 9 retries takes) and raising an
EcsEventualConsistencyTimeout if we exhaust our retries. This might be
playing with fire but I wanted to balance the waiting against the
expectation that a new run should launch relatively close to when
Dagster requests it.