[demo pipelines] remove dynamic download pipeline
Summary:
The goal of the demo repo is to simulate what a "real" set of data application would look like, and a real data application would not include two parallel implementations of the same functionality.  I think there are many opportunities for confusion if someone is trying to copy/paste code or figure out what patterns they themselves should follow.

If we want to showcase dynamic orchestration, we could add an additional pipeline or set of steps that require it.  Maybe some hyperparameter search for the story recommender or something.

Test Plan: bk

Reviewers: owen

Reviewed By: owen

Differential Revision: https://dagster.phacility.com/D9215