Allow empty IWorkspace to be passed in to a LaunchRunContext
Summary:
The only run launcher that requires a workspace is DefaultRunLauncher. Letting the workspace field be empty allows run launchers to be used in other environments that won't ever use a DefaultRunLauncher.

(Also remove some back-compat code that is no longer needed now that we are past the 0.12.0 threshold)

Test Plan: BK