improve Airflow integration docs #4528
The Airflow integration docs were confusingly spread across two different places, which made many readers believe that we either only supported Airflow ingestion or Airflow export not both.

They were originally part of the intro tutorial as far as I can tell, which leads to the main points kind of getting lost in a sea of explaining Docker and explaining Airflow. That content made sense when Airflow was the recommended way of deploying Dagster, but makes less sense where the main scenario for coming to the integration is already having an existing functioning Airflow setup. So this version cuts that stuff out.

Fixes #3891