Fix docker run launcher with no network or networks set
Summary:
This was erroring out in launch_run. Testing is a little funky since we kind of need it to be set for the run to actually complete (a mocked out docker client would help here), but we can still validate that the launch_run call completes successfully and a container is started.

Test Plan: BK