Raise a more useful error when you don't implement the static/class methods on configurable class (#4836)
Summary:
You would think you wouldn't need this because they're static/class methods, but you would be wrong :) The instance ref loading path instantiates this class and is able to call from_config_value on it even when the method is not implemented! It will fail later, but much more cryptically, so add a clearer error message.