Initialize sub_process in execute_script_file
Summary:
If the POpen constructor raises an exception for some reason, the variable is not initialized.

Test Plan: I tried for a while to come up with a situation where this would happen and was not successful, but https://dagster.slack.com/archives/C01U954MEER/p1632425935208000 suggests it is happening somehow some of the time.