RFC: Helm template to allow specifying volumes in camelcase in
Summary:
This is some truly hairy helm code that i'm not sure that we should land, but it recursively goes through volumeMounts and transforms the keys in any dict that it finds into camel case. This would allow us to specify volumeMounts and volumes config in camel case and make use of the built-in k8s helm validation schemas.

However, in addition to just being really gross, it also will incorrectly snake case any fields that are dicts with user-specified keys (which are rare, but do exist)