Don't allow grpc servers to fail due to heartbeat failure before the startup code times out
Summary:
Some recent changes to the gRPC server startup have produced strange/confusing error behavior when the server starts up correctly, but is not reachable from dagit (for example, because it's behind a VPN, or takes a super long time to load). Because our heartbeat timeout is shorter than our tiemout, the first indication they see that something goes wrong is 'the gRPC server exited with error code 0'. This is confusing.

Solve this in two ways:
- ensure that we timeout before the hearbeat expires (by bumping up the hearteat a bit, and ensuring the timeout is always less than the heartbeat TTL)
- include the test of the most recent connection error when it timesout. This helps with things like the VPN case where the error gives some clues why it might be timing out.