Yield in the correct order
Originally, these tests continued to pass against the old implementation
(before https://github.com/dagster-io/dagster/pull/5184 fixed the key
lookup behavior).

It turns out we were yielding in the wrong order in our test case. When
we yield `foo` then `foobar`, both get written under either
implementation because `foobar` is not a prefix of `foo`.

When we flip the test case to yield `foobar` and then `foo`, the old
implementation fails because `foo` is a prefix of `foobar` - the `foo`
output overwrites the `foobar` output.