Fixup prefixed IO manager tests (#5189)
This test was failing initially because of a syntax error. But once I
fixed that, it continued to fail because no outputs were actually being
written.

I switched from the existing pattern of building and traversing the
execution plan by hand to our new `job.execute_in_process()` method and
it helped pinpoint the issue to us not yielding an `Output()` from our
job.

I like this test implementation because I think it makes the behavior
under test easier to understand - we expect to see two keys written to
S3 and we expect them to contain the appropriate outputs.