Fix error in re-execution of a dynamic pipeline (#5248)
Summary:
This happened when I tried to run a pipeline with lots of dynamic outputs by a user with lots of dynamic outputs. This was calling a method that did not exist on the collect version of this class.

Based on https://dagster.phacility.com/D7289#change-DsMVnB8AgF1e this seems to have been the intention (and the error I was seeing no longer happens)

Test Plan: Test test_reexec_from_parent_dynamic_fails still succeeds. I do not quite have the dynamic pipeline wherewithal to identify a test case that covers the collect case though...