Use executable path when making execute_run and execute_step calls
Summary:
A couple of different users lately have asked for the ability to have a single Docker image with multiple venvs in it, using different grpc server entry points in the same image. We are shockingly close to being able to support that - the executable path is even passed all the way through, but then we don't use it when spinning up the pods / containers /etc. - this changes that by including the executable path in the command.

Test Plan: BK