Don't mark a job as succeeded when an op calls sys.exit
Summary:
Noticed this when trying to test op crashes - this finally block incorrectly marks the pipeline as succeeded when a BaseException is raised that isn't an Exception (and isn't a KeyboardInterrupt).

The chart on https://docs.python.org/3/library/exceptions.html#exception-hierarchy seems to indicate that SystemException is the only class that this affects.

Test Plan: BK