Reduce dependencies that bring in regex
Summary:
regex regularly pushes and breaks our builds while we wait for the wheels to build. We don't really need it in dagster[test], and its brought in by a docker validation library that only gives us a marginally better error message when a malformed docker image is used. Remove both of those deps.

Sadly papermill still brings in black and thus regex, but dagstermill is included in many fewer builds.