RFC: Set a fixed pip version when running tox suites (#5530)
Summary:
Right now when you run a tox suite, it uses the default virtualenv version that's installed in the automation library, which is 19. We could increase that version, but it's probably good to be able to rely that all our tox suites are running on the same version. Luckily, there is a tox flag that provides this.