Don't set ECS tags if we're using short ARNs (#5651)
ECS can only use tags if the account is using the new long ARN format.

https://docs.aws.amazon.com/AmazonECS/latest/developerguide/ecs-using-tags.html
https://docs.aws.amazon.com/AmazonECS/latest/developerguide/ecs-account-settings.html#ecs-resource-ids

Previously, we were adding a tag to the ECS task with the id of the
Dagster run. This drove no actual behavior - it's purely informational
metadata.

Now, we'll skip applying the tag if your account isn't using the new
long ARN format.