Fix issue where we were over-requiring resources
Summary:
When op subsetting, we were adding all resources that any resource in the job depended on (instead of just the resources in the op being subsetted). Not sure if there's additional filtering that needs to happen in the other cases in this function to account for that, but this is the one that I ran into. Will add a test that simulates the issue in the release pipeline.