Don't fail runs when they are force-terminated
Summary:
I noticed that when you force-cancel a run, you get a strange/confusing message: "This pipeline was forcibly marked as canceled from outside the execution context. The computational resources created by the run may not have been fully cleaned up." which doesn't make sense in this context, and also results in the run getting moved into FAILED instead of CANCELED. You might also end up resuming the run incorrectly after force-terminating it.

Instead, add a case that specifically handles the CANCELLED-but-then-the-run-still-cleans-up-nicely case, and leaves it in CANCELED.