Allow json serialized strings to be passed in as run config to graphql calls
Summary:
In some languages / graphql clients it is a pain to set up a scalar that is a JSON object, and it's easier for clients to serialize up the run config as a string and upload that (even if it ends up double-encoded since the transport is also JSON-encoded).

This PR adds that as an option.