Use the `dagster` CLI entry point by default when launching dagster runs and steps via CLI, instead of <python_path> -m dagster api grpc
Summary:
In https://github.com/dagster-io/dagster/pull/5415 we started using the executable_path much more frequently when running commands - it basically became the default in all environments to pull sys.executable from the gRPC server and then use that to determine the entry point, even in environemnts like Docker/K8s/etc. where using `dagster` as the entry point was working fine. That change was to support a couple of different users who reported running mutliple venvs in a single Docker image, and having no way to tell Dagster to use that venv on run launch.

However, that change seems to have, perhaps predictably in hindsight, affected another class of users, who run a single Dagster executable/entrypoint and want to route all dagster commands through that (i.e assuming that sys.executable points at a Python runtime specifically that can be called with -m is not a valid assumption in general). Moreover, you can imagine people doing things like running dagster in custom images not neccesarily having the exact same python entry point. So making the change above in a blanket fashion may not have been the best move.

To support both use cases, this is a partial revert of #5415 - instead of being the default behavior, it's now something you can explicitly enable with a flag on the gRPC server - this will then make the executable_path non-null, which will ensure that that executable path. In the common case though, we switch back to the deafutl

This will be a bit of a breaking change for those couple of users who reported the multiple venv behavior originally, but I'll make sure they are aware of that before this goes out, assumign this seems like a good idea..