Make connect_container_to_network idempotent (#6137)
On Buildkite, if you try to use the same network with multiple docker
compose files, the Buildkite test container can't connect to the network
a second time. `connect_container_to_network` returns a returncode of 1
and the Docker daemon emits:

```
Error response from daemon: endpoint with name {container} already exists in network {network}
```

Now we just continue on if we get an error while trying to connect.
There's probably a more sophisticated approach to this (for example,
checking to see if the network is already attached before trying to
attach it) but I think this will be sufficient for now.