Jobify dagster new-project
Summary:
ran into this when I was trying to clean up refs to python_package/python_module in the docs. Docs still reference solids, code still defaults to graph, and IMO this is much to early to drop the confusing 'repository location' term.

Test Plan: BK, gen a new module, load it in dagit