simple test case for working dir issue
Summary:
Based on the description in slack I'd expect this to fail, but for some reason it is not:

docker build . -t test
docker run test

Seems to load the code just fine.

(But if I add a --empty-working-directory arg to the gRPC server, then it gives an error about not being able to load the module).