Make runs tolerant to context.log machinery errors for user-generated logs
Summary:
This is intended to be a lot like @prha's https://github.com/dagster-io/dagster/pull/6122, but for log.info calls instead of compute log calls. If you write a log.info call and it fails, that's worth logging, but probably is not enough to justify failing the entire step and run.

Specifically really large log calls might run into size limits or other issues related to the content of the custom log, and that shouldn't fail the whole run.

Contrast with failures writing the system events, which do meet the bar for failing the run since they will gum up the execution as a whole)