RFC: Only run Bk on all python versions on master/release branches
Summary:
Adding 3.9 (and soon 3.10) is good, but the sheer number of BK steps running on every PR is getting pretty overwhelming and hard to parse, and actual version differences are quite rare in my experience (to the point where I'd personally be fine waiting to see them in master). There's probably a cost control benefit here as well, but that wasn't the main reason I was inspired to send this out.

None of this is held strongly.