Reconcile AWS deployment docs with ECS deployment docs (#6526)
Summary:
Right now there's an ECS-specific guide but the only way to find it is to link through from the EcsRunLauncher. This attempts to reconcile them by linking directly to ECS from AWS and only having the example live in ECS.

I could also be convinced to put these both in the AWS page and just make it kind of long.