increase truncation threshold on jobs page (#6583)
Summary:
This is related to https://github.com/dagster-io/dagster/pull/6286 - which changed the truncation threshold to jobs to 40 characters. That makes sense on the RunRoot page that prompted the PR, but this component is also used in the jobs page where there is a lot more horizontal space (a user reported that their long job names were now indistinguishable)

That said I'm not sure using a fixed truncation limit in that component makes sense - maybe it should be controlled in CSS in that callsite? When I tried taking out the limit altogether, it didn't seem like there was any limit and the table just expanded undesirably.

Test Plan: Load jobs table with a really ong job, it is no longer being truncated prematurely