Default helm chart to enable the QueuedRunCoordinator (#6551)
For 0.14.0

A @alangenfeld pointed out, either choice here has the potential to surprise users. Enabling it means they can hit the 10 run default limit, while disabling it means they can accidentally overwhelm their cluster (though it's probably possible for 10 runs to do this to a small cluster anyway).

My take is that we're better at surfacing (and therefore making it easier for users to fix) that runs are stuck in the queue (you'll see it on the runs page, with a link to the queue configuration) versus surfacing that your cluster is getting overwhelmed.