Order get_ticks by a column that we have an index on in JobTickTable
Summary:
A user reported a timeout on this query when there are large numbers of ticks. We have an index on job_origin_id + timestamp, but not on job_origin_id + id.

I'm a little confused about what primary keys are good for as indices when there are other filters applied - would like to discuss if https://github.com/dagster-io/internal/pull/1139 is actually sufficient to make a similar query more efficient?

Test Plan: Verify with a postgres query plan that the get_ticks query is now performant